VAR: Estimation

Estimate VAR(1) models by efficiently sampling from the posterior distribution. This provides two graphical structures: (1) a network of undirected relations (the GGM, controlling for the lagged predictors) and (2) a network of directed relations (the lagged coefficients). Note that in the graphical modeling literature, this model is also known as a time series chain graphical model (Abegaz and Wit 2013) .

var_estimate(
  Y,
  rho_sd = 0.5,
  beta_sd = 1,
  iter = 5000,
  progress = TRUE,
  seed = 1,
  ...
)

tscgm_estimate(
  Y,
  rho_sd = 0.5,
  beta_sd = 1,
  iter = 5000,
  progress = TRUE,
  seed = 1,
  ...
)

Arguments

Y
Matrix (or data frame) of dimensions n (observations) by p (variables).

rho_sd
Numeric. Scale of the prior distribution for the partial correlations, approximately the standard deviation of a beta distribution (defaults to 0.50).

beta_sd
Numeric. Standard deviation of the prior distribution for the regression coefficients (defaults to 1). The prior is by default centered at zero and follows a normal distribution (Equation 9, Sinay and Hsu 2014)

iter
Number of iterations (posterior samples; defaults to 5000).

progress
Logical. Should a progress bar be included (defaults to TRUE) ?

seed
An integer for the random seed (defaults to 1).

...
Currently ignored.

Value

An object of class var_estimate containing a lot of information that is used for printing and plotting the results. For users of BGGM, the following are the useful objects:

beta_mu A matrix including the regression coefficients (posterior mean).

pcor_mu Partial correlation matrix (posterior mean).

fit A list including the posterior samples.

Details

Each time series in Y is standardized (mean = 0; standard deviation = 1).

Note

Regularization:

A Bayesian ridge regression can be fitted by decreasing beta_sd (e.g., beta_sd = 0.25). This could be advantageous for forecasting (out-of-sample prediction) in particular.

References

Abegaz F, Wit E (2013). “Sparse time series chain graphical models for reconstructing genetic networks.” Biostatistics, 14(3), 586--599.

Sinay MS, Hsu JS (2014). “Bayesian inference of a multivariate regression model.” Journal of Probability and Statistics, 2014.

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# use alias (var_estimate also works)
fit <- tscgm_estimate(Y, progress = FALSE)

#> BGGM: Posterior Sampling 

#> BGGM: Finished

fit

#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Posterior Samples: 5000 
#> Observations (n): 94 
#> Nodes (p): 7 
#> --- 
#> Call: 
#> tscgm_estimate(Y = Y, progress = FALSE)
#> --- 
#> Partial Correlations: 
#> 
#>               interested disinterested excited  upset strong stressed  steps
#> interested         0.000        -0.186   0.378 -0.196  0.328    0.275  0.074
#> disinterested     -0.186         0.000  -0.169 -0.024  0.102    0.162 -0.079
#> excited            0.378        -0.169   0.000 -0.134  0.493   -0.175 -0.008
#> upset             -0.196        -0.024  -0.134  0.000  0.113    0.344 -0.038
#> strong             0.328         0.102   0.493  0.113  0.000   -0.013  0.184
#> stressed           0.275         0.162  -0.175  0.344 -0.013    0.000 -0.011
#> steps              0.074        -0.079  -0.008 -0.038  0.184   -0.011  0.000
#> --- 
#> Coefficients: 
#> 
#>                  interested disinterested excited  upset strong stressed  steps
#> interested.l1         0.220        -0.013   0.176 -0.096  0.171    0.015  0.106
#> disinterested.l1     -0.050        -0.004   0.056 -0.018  0.049    0.089 -0.023
#> excited.l1           -0.084        -0.184  -0.001  0.055 -0.085    0.085  0.099
#> upset.l1             -0.153         0.256  -0.099  0.432  0.054    0.317 -0.095
#> strong.l1             0.029         0.173   0.030  0.046  0.185   -0.065 -0.182
#> stressed.l1          -0.020        -0.009  -0.029 -0.044 -0.073    0.150  0.133
#> steps.l1             -0.153         0.181  -0.205  0.150 -0.089    0.205  0.039
#> --- 
#> Date: Sat Jun 06 09:17:55 2020 

# }