Data: 25 Personality items representing 5 factors

This dataset and the corresponding documentation was taken from the psych package. We refer users to that package for further details (Revelle 2019) .

data("bfi")

Format

A data frame with 25 variables and 2800 observations (including missing values)

Details

A1 Am indifferent to the feelings of others. (q_146)

A2 Inquire about others' well-being. (q_1162)

A3 Know how to comfort others. (q_1206)

A4 Love children. (q_1364)

A5 Make people feel at ease. (q_1419)

C1 Am exacting in my work. (q_124)

C2 Continue until everything is perfect. (q_530)

C3 Do things according to a plan. (q_619)

C4 Do things in a half-way manner. (q_626)

C5 Waste my time. (q_1949)

E1 Don't talk a lot. (q_712)

E2 Find it difficult to approach others. (q_901)

E3 Know how to captivate people. (q_1205)

E4 Make friends easily. (q_1410)

E5 Take charge. (q_1768)

N1 Get angry easily. (q_952)

N2 Get irritated easily. (q_974)

N3 Have frequent mood swings. (q_1099)

N4 Often feel blue. (q_1479)

N5 Panic easily. (q_1505)

o1 Am full of ideas. (q_128)

o2 Avoid difficult reading material.(q_316)

o3 Carry the conversation to a higher level. (q_492)

o4 Spend time reflecting on things. (q_1738)

o5 Will not probe deeply into a subject. (q_1964)

gender Males = 1, Females =2

education 1 = HS, 2 = finished HS, 3 = some college, 4 = college graduate 5 = graduate degree

References

Revelle W (2019). psych: Procedures for Psychological, Psychometric, and Personality Research. Northwestern University, Evanston, Illinois. R package version 1.9.12, https://CRAN.R-project.org/package=psych.