Compute Regression Parameters for explore Objects

There is a direct correspondence between the inverse covariance matrix and multiple regression (Kwan 2014; Stephens 1998) . This readily allows for converting the GGM paramters to regression coefficients. All data types are supported.

# S3 method for explore
coef(object, iter = NULL, ...)

Arguments

object
An Object of class explore.

iter
Number of iterations (posterior samples; defaults to the number in the object).

...
Currently ignored.

Value

An object of class coef, containting two lists.

betas A list of length p, each containing a p - 1 by iter matrix of posterior samples

object An object of class explore (the fitted model).

References

Kwan CC (2014). “A regression-based interpretation of the inverse of the sample covariance matrix.” Spreadsheets in Education, 7(1), 4613.

Stephens G (1998). “On the Inverse of the Covariance Matrix in Portfolio Analysis.” The Journal of Finance, 53(5), 1821--1827.

Examples

# \donttest{
# note: iter = 250 for demonstrative purposes

data

#> function (..., list = character(), package = NULL, lib.loc = NULL, 
#>     verbose = getOption("verbose"), envir = .GlobalEnv, overwrite = TRUE) 
#> {
#>     fileExt <- function(x) {
#>         db <- grepl("\\.[^.]+\\.(gz|bz2|xz)$", x)
#>         ans <- sub(".*\\.", "", x)
#>         ans[db] <- sub(".*\\.([^.]+\\.)(gz|bz2|xz)$", "\\1\\2", 
#>             x[db])
#>         ans
#>     }
#>     my_read_table <- function(...) {
#>         lcc <- Sys.setlocale("LC_COLLATE", "C")
#>         on.exit(Sys.setlocale("LC_COLLATE", lcc))
#>         read.table(...)
#>     }
#>     names <- c(as.character(substitute(list(...))[-1L]), list)
#>     if (!is.null(package)) {
#>         if (!is.character(package)) 
#>             stop("'package' must be a character string or NULL")
#>         if (FALSE) {
#>             if (any(package %in% "base")) 
#>                 warning("datasets have been moved from package 'base' to package 'datasets'")
#>             if (any(package %in% "stats")) 
#>                 warning("datasets have been moved from package 'stats' to package 'datasets'")
#>             package[package %in% c("base", "stats")] <- "datasets"
#>         }
#>     }
#>     paths <- find.package(package, lib.loc, verbose = verbose)
#>     if (is.null(lib.loc)) 
#>         paths <- c(path.package(package, TRUE), if (!length(package)) getwd(), 
#>             paths)
#>     paths <- unique(normalizePath(paths[file.exists(paths)]))
#>     paths <- paths[dir.exists(file.path(paths, "data"))]
#>     dataExts <- tools:::.make_file_exts("data")
#>     if (length(names) == 0L) {
#>         db <- matrix(character(), nrow = 0L, ncol = 4L)
#>         for (path in paths) {
#>             entries <- NULL
#>             packageName <- if (file_test("-f", file.path(path, 
#>                 "DESCRIPTION"))) 
#>                 basename(path)
#>             else "."
#>             if (file_test("-f", INDEX <- file.path(path, "Meta", 
#>                 "data.rds"))) {
#>                 entries <- readRDS(INDEX)
#>             }
#>             else {
#>                 dataDir <- file.path(path, "data")
#>                 entries <- tools::list_files_with_type(dataDir, 
#>                   "data")
#>                 if (length(entries)) {
#>                   entries <- unique(tools::file_path_sans_ext(basename(entries)))
#>                   entries <- cbind(entries, "")
#>                 }
#>             }
#>             if (NROW(entries)) {
#>                 if (is.matrix(entries) && ncol(entries) == 2L) 
#>                   db <- rbind(db, cbind(packageName, dirname(path), 
#>                     entries))
#>                 else warning(gettextf("data index for package %s is invalid and will be ignored", 
#>                   sQuote(packageName)), domain = NA, call. = FALSE)
#>             }
#>         }
#>         colnames(db) <- c("Package", "LibPath", "Item", "Title")
#>         footer <- if (missing(package)) 
#>             paste0("Use ", sQuote(paste("data(package =", ".packages(all.available = TRUE))")), 
#>                 "\n", "to list the data sets in all *available* packages.")
#>         else NULL
#>         y <- list(title = "Data sets", header = NULL, results = db, 
#>             footer = footer)
#>         class(y) <- "packageIQR"
#>         return(y)
#>     }
#>     paths <- file.path(paths, "data")
#>     for (name in names) {
#>         found <- FALSE
#>         for (p in paths) {
#>             tmp_env <- if (overwrite) 
#>                 envir
#>             else new.env()
#>             if (file_test("-f", file.path(p, "Rdata.rds"))) {
#>                 rds <- readRDS(file.path(p, "Rdata.rds"))
#>                 if (name %in% names(rds)) {
#>                   found <- TRUE
#>                   if (verbose) 
#>                     message(sprintf("name=%s:\t found in Rdata.rds", 
#>                       name), domain = NA)
#>                   thispkg <- sub(".*/([^/]*)/data$", "\\1", p)
#>                   thispkg <- sub("_.*$", "", thispkg)
#>                   thispkg <- paste0("package:", thispkg)
#>                   objs <- rds[[name]]
#>                   lazyLoad(file.path(p, "Rdata"), envir = tmp_env, 
#>                     filter = function(x) x %in% objs)
#>                   break
#>                 }
#>                 else if (verbose) 
#>                   message(sprintf("name=%s:\t NOT found in names() of Rdata.rds, i.e.,\n\t%s\n", 
#>                     name, paste(names(rds), collapse = ",")), 
#>                     domain = NA)
#>             }
#>             if (file_test("-f", file.path(p, "Rdata.zip"))) {
#>                 warning("zipped data found for package ", sQuote(basename(dirname(p))), 
#>                   ".\nThat is defunct, so please re-install the package.", 
#>                   domain = NA)
#>                 if (file_test("-f", fp <- file.path(p, "filelist"))) 
#>                   files <- file.path(p, scan(fp, what = "", quiet = TRUE))
#>                 else {
#>                   warning(gettextf("file 'filelist' is missing for directory %s", 
#>                     sQuote(p)), domain = NA)
#>                   next
#>                 }
#>             }
#>             else {
#>                 files <- list.files(p, full.names = TRUE)
#>             }
#>             files <- files[grep(name, files, fixed = TRUE)]
#>             if (length(files) > 1L) {
#>                 o <- match(fileExt(files), dataExts, nomatch = 100L)
#>                 paths0 <- dirname(files)
#>                 paths0 <- factor(paths0, levels = unique(paths0))
#>                 files <- files[order(paths0, o)]
#>             }
#>             if (length(files)) {
#>                 for (file in files) {
#>                   if (verbose) 
#>                     message("name=", name, ":\t file= ...", .Platform$file.sep, 
#>                       basename(file), "::\t", appendLF = FALSE, 
#>                       domain = NA)
#>                   ext <- fileExt(file)
#>                   if (basename(file) != paste0(name, ".", ext)) 
#>                     found <- FALSE
#>                   else {
#>                     found <- TRUE
#>                     zfile <- file
#>                     zipname <- file.path(dirname(file), "Rdata.zip")
#>                     if (file.exists(zipname)) {
#>                       Rdatadir <- tempfile("Rdata")
#>                       dir.create(Rdatadir, showWarnings = FALSE)
#>                       topic <- basename(file)
#>                       rc <- .External(C_unzip, zipname, topic, 
#>                         Rdatadir, FALSE, TRUE, FALSE, FALSE)
#>                       if (rc == 0L) 
#>                         zfile <- file.path(Rdatadir, topic)
#>                     }
#>                     if (zfile != file) 
#>                       on.exit(unlink(zfile))
#>                     switch(ext, R = , r = {
#>                       library("utils")
#>                       sys.source(zfile, chdir = TRUE, envir = tmp_env)
#>                     }, RData = , rdata = , rda = load(zfile, 
#>                       envir = tmp_env), TXT = , txt = , tab = , 
#>                       tab.gz = , tab.bz2 = , tab.xz = , txt.gz = , 
#>                       txt.bz2 = , txt.xz = assign(name, my_read_table(zfile, 
#>                         header = TRUE, as.is = FALSE), envir = tmp_env), 
#>                       CSV = , csv = , csv.gz = , csv.bz2 = , 
#>                       csv.xz = assign(name, my_read_table(zfile, 
#>                         header = TRUE, sep = ";", as.is = FALSE), 
#>                         envir = tmp_env), found <- FALSE)
#>                   }
#>                   if (found) 
#>                     break
#>                 }
#>                 if (verbose) 
#>                   message(if (!found) 
#>                     "*NOT* ", "found", domain = NA)
#>             }
#>             if (found) 
#>                 break
#>         }
#>         if (!found) {
#>             warning(gettextf("data set %s not found", sQuote(name)), 
#>                 domain = NA)
#>         }
#>         else if (!overwrite) {
#>             for (o in ls(envir = tmp_env, all.names = TRUE)) {
#>                 if (exists(o, envir = envir, inherits = FALSE)) 
#>                   warning(gettextf("an object named %s already exists and will not be overwritten", 
#>                     sQuote(o)))
#>                 else assign(o, get(o, envir = tmp_env, inherits = FALSE), 
#>                   envir = envir)
#>             }
#>             rm(tmp_env)
#>         }
#>     }
#>     invisible(names)
#> }
#> <bytecode: 0x000000002cef29b8>
#> <environment: namespace:utils>

Y <- ptsd
#########################################
### example 1: continuous and ordinal ###
#########################################

# fit model
fit <- explore(Y, type = "continuous",
               iter = 250)

#> BGGM: Posterior Sampling 

#> BGGM: Prior Sampling 

#> BGGM: Finished


# regression
reg <- coef(fit)

#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |==========                                                            |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |========================                                              |  35%  |                                                                              |============================                                          |  40%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |====================================================                  |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |==================================================================    |  95%  |                                                                              |======================================================================| 100%


# summary
summ <- summary(reg)

summ

#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Type: continuous 
#> Formula:  
#> --- 
#> Call: 
#> explore(Y = Y, type = "continuous", iter = 250)
#> --- 
#> Coefficients: 
#>  
#> B1: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B2     0.206   0.077   0.045   0.367
#>    B3     0.022   0.086  -0.133   0.189
#>    B4     0.338   0.072   0.196   0.480
#>    B5     0.147   0.072   0.003   0.263
#>    C1    -0.033   0.059  -0.146   0.080
#>    C2    -0.049   0.067  -0.180   0.070
#>    D1     0.051   0.057  -0.058   0.160
#>    D2    -0.037   0.063  -0.149   0.086
#>    D3    -0.070   0.072  -0.224   0.068
#>    D4     0.225   0.080   0.080   0.391
#>    D5     0.090   0.066  -0.020   0.220
#>    D6     0.061   0.076  -0.077   0.219
#>    D7     0.034   0.075  -0.101   0.179
#>    E1     0.051   0.068  -0.089   0.179
#>    E2    -0.019   0.061  -0.146   0.096
#>    E3    -0.031   0.065  -0.153   0.093
#>    E4    -0.151   0.069  -0.287  -0.023
#>    E5     0.035   0.069  -0.088   0.175
#>    E6     0.046   0.068  -0.086   0.189
#> 
#> B2: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.192   0.072   0.039   0.324
#>    B3     0.407   0.066   0.290   0.526
#>    B4    -0.030   0.075  -0.179   0.114
#>    B5     0.041   0.072  -0.091   0.178
#>    C1    -0.021   0.059  -0.149   0.086
#>    C2     0.081   0.065  -0.062   0.216
#>    D1    -0.006   0.054  -0.114   0.106
#>    D2    -0.089   0.058  -0.203   0.031
#>    D3     0.036   0.066  -0.091   0.157
#>    D4     0.058   0.076  -0.088   0.229
#>    D5     0.081   0.071  -0.061   0.198
#>    D6     0.053   0.075  -0.102   0.203
#>    D7    -0.095   0.072  -0.244   0.032
#>    E1    -0.072   0.070  -0.213   0.057
#>    E2     0.144   0.061   0.025   0.268
#>    E3    -0.014   0.054  -0.118   0.089
#>    E4     0.100   0.066  -0.036   0.215
#>    E5    -0.066   0.069  -0.208   0.054
#>    E6     0.120   0.056   0.021   0.226
#> 
#> B3: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.019   0.074  -0.112   0.168
#>    B2     0.376   0.062   0.268   0.500
#>    B4     0.263   0.070   0.134   0.396
#>    B5     0.094   0.066  -0.041   0.219
#>    C1    -0.025   0.055  -0.128   0.090
#>    C2    -0.001   0.060  -0.113   0.121
#>    D1    -0.012   0.053  -0.120   0.088
#>    D2    -0.106   0.060  -0.225   0.016
#>    D3     0.020   0.064  -0.107   0.141
#>    D4     0.025   0.073  -0.107   0.167
#>    D5     0.110   0.067  -0.020   0.241
#>    D6    -0.058   0.062  -0.159   0.066
#>    D7     0.096   0.063  -0.024   0.214
#>    E1     0.010   0.063  -0.104   0.146
#>    E2     0.000   0.055  -0.113   0.097
#>    E3     0.089   0.056  -0.017   0.193
#>    E4     0.051   0.068  -0.075   0.173
#>    E5     0.056   0.065  -0.075   0.178
#>    E6     0.041   0.055  -0.071   0.143
#> 
#> B4: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.297   0.063   0.181   0.413
#>    B2    -0.028   0.070  -0.165   0.107
#>    B3     0.270   0.070   0.119   0.410
#>    B5     0.206   0.069   0.080   0.348
#>    C1     0.096   0.057  -0.014   0.198
#>    C2     0.133   0.064   0.018   0.260
#>    D1    -0.014   0.052  -0.112   0.083
#>    D2     0.162   0.063   0.023   0.282
#>    D3     0.067   0.065  -0.053   0.188
#>    D4     0.066   0.076  -0.061   0.219
#>    D5    -0.106   0.066  -0.234   0.019
#>    D6    -0.086   0.072  -0.239   0.046
#>    D7    -0.104   0.066  -0.241   0.014
#>    E1    -0.003   0.068  -0.123   0.133
#>    E2    -0.040   0.055  -0.144   0.053
#>    E3    -0.002   0.059  -0.121   0.103
#>    E4     0.033   0.075  -0.118   0.182
#>    E5    -0.052   0.065  -0.189   0.075
#>    E6    -0.002   0.058  -0.122   0.105
#> 
#> B5: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.140   0.069   0.002   0.263
#>    B2     0.042   0.072  -0.087   0.169
#>    B3     0.104   0.074  -0.046   0.243
#>    B4     0.222   0.073   0.079   0.364
#>    C1     0.152   0.056   0.033   0.255
#>    C2     0.121   0.067   0.001   0.248
#>    D1    -0.063   0.055  -0.172   0.041
#>    D2    -0.047   0.067  -0.188   0.071
#>    D3     0.071   0.068  -0.054   0.198
#>    D4    -0.007   0.072  -0.147   0.118
#>    D5    -0.023   0.070  -0.160   0.122
#>    D6    -0.054   0.076  -0.196   0.086
#>    D7    -0.020   0.074  -0.164   0.112
#>    E1     0.177   0.069   0.054   0.324
#>    E2    -0.038   0.056  -0.146   0.075
#>    E3    -0.037   0.062  -0.161   0.088
#>    E4     0.173   0.069   0.049   0.296
#>    E5     0.040   0.072  -0.118   0.168
#>    E6     0.063   0.060  -0.052   0.169
#> 
#> C1: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.049   0.090  -0.222   0.122
#>    B2    -0.033   0.096  -0.227   0.137
#>    B3    -0.045   0.097  -0.224   0.138
#>    B4     0.165   0.097  -0.026   0.343
#>    B5     0.241   0.087   0.057   0.396
#>    C2     0.223   0.078   0.068   0.369
#>    D1     0.105   0.061  -0.019   0.215
#>    D2    -0.149   0.082  -0.299   0.014
#>    D3     0.117   0.082  -0.041   0.280
#>    D4    -0.090   0.101  -0.286   0.098
#>    D5     0.001   0.087  -0.140   0.186
#>    D6     0.116   0.095  -0.074   0.299
#>    D7     0.034   0.088  -0.120   0.199
#>    E1    -0.046   0.090  -0.218   0.128
#>    E2     0.002   0.079  -0.161   0.141
#>    E3     0.118   0.075  -0.034   0.254
#>    E4    -0.046   0.087  -0.219   0.109
#>    E5    -0.013   0.083  -0.162   0.162
#>    E6    -0.021   0.076  -0.158   0.122
#> 
#> C2: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.061   0.083  -0.223   0.077
#>    B2     0.108   0.086  -0.083   0.274
#>    B3    -0.001   0.085  -0.162   0.158
#>    B4     0.185   0.085   0.019   0.349
#>    B5     0.158   0.087   0.001   0.339
#>    C1     0.182   0.064   0.056   0.305
#>    D1     0.067   0.060  -0.053   0.188
#>    D2    -0.033   0.084  -0.222   0.133
#>    D3     0.006   0.078  -0.157   0.140
#>    D4     0.003   0.087  -0.191   0.168
#>    D5     0.215   0.068   0.066   0.328
#>    D6    -0.012   0.096  -0.205   0.171
#>    D7     0.098   0.081  -0.052   0.271
#>    E1     0.126   0.069  -0.010   0.244
#>    E2     0.042   0.068  -0.082   0.170
#>    E3     0.021   0.067  -0.124   0.138
#>    E4     0.047   0.079  -0.119   0.209
#>    E5    -0.169   0.075  -0.308  -0.032
#>    E6    -0.066   0.071  -0.221   0.059
#> 
#> D1: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.086   0.094  -0.094   0.253
#>    B2    -0.011   0.099  -0.205   0.186
#>    B3    -0.023   0.104  -0.241   0.187
#>    B4    -0.028   0.101  -0.229   0.168
#>    B5    -0.116   0.102  -0.332   0.070
#>    C1     0.119   0.070  -0.023   0.242
#>    C2     0.094   0.084  -0.070   0.270
#>    D2     0.069   0.088  -0.093   0.236
#>    D3     0.120   0.090  -0.054   0.307
#>    D4     0.012   0.106  -0.197   0.226
#>    D5    -0.024   0.088  -0.186   0.170
#>    D6     0.004   0.099  -0.170   0.195
#>    D7    -0.070   0.096  -0.252   0.115
#>    E1     0.040   0.087  -0.144   0.182
#>    E2     0.124   0.076  -0.019   0.283
#>    E3     0.109   0.081  -0.069   0.256
#>    E4    -0.033   0.099  -0.207   0.152
#>    E5     0.177   0.087  -0.002   0.330
#>    E6    -0.037   0.074  -0.155   0.113
#> 
#> D2: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.042   0.072  -0.169   0.103
#>    B2    -0.111   0.071  -0.247   0.042
#>    B3    -0.144   0.082  -0.292   0.023
#>    B4     0.216   0.086   0.036   0.369
#>    B5    -0.056   0.081  -0.214   0.087
#>    C1    -0.116   0.066  -0.240   0.010
#>    C2    -0.031   0.080  -0.204   0.125
#>    D1     0.048   0.060  -0.065   0.165
#>    D3     0.096   0.083  -0.073   0.256
#>    D4     0.268   0.082   0.116   0.439
#>    D5     0.224   0.074   0.068   0.364
#>    D6     0.076   0.080  -0.088   0.223
#>    D7     0.171   0.077   0.020   0.317
#>    E1     0.117   0.065  -0.007   0.239
#>    E2    -0.035   0.064  -0.160   0.083
#>    E3     0.099   0.069  -0.036   0.233
#>    E4    -0.017   0.076  -0.162   0.144
#>    E5     0.046   0.072  -0.103   0.171
#>    E6    -0.021   0.068  -0.153   0.112
#> 
#> D3: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.076   0.079  -0.234   0.073
#>    B2     0.044   0.081  -0.119   0.193
#>    B3     0.025   0.081  -0.130   0.185
#>    B4     0.084   0.081  -0.069   0.236
#>    B5     0.082   0.078  -0.061   0.228
#>    C1     0.085   0.060  -0.031   0.196
#>    C2     0.005   0.071  -0.140   0.127
#>    D1     0.077   0.058  -0.031   0.188
#>    D2     0.090   0.078  -0.065   0.233
#>    D4     0.563   0.070   0.431   0.702
#>    D5    -0.145   0.076  -0.292   0.002
#>    D6    -0.002   0.083  -0.162   0.166
#>    D7    -0.051   0.074  -0.194   0.084
#>    E1    -0.078   0.071  -0.229   0.072
#>    E2     0.111   0.065  -0.009   0.231
#>    E3    -0.043   0.060  -0.177   0.075
#>    E4    -0.133   0.071  -0.255   0.014
#>    E5     0.106   0.072  -0.048   0.231
#>    E6     0.033   0.059  -0.074   0.140
#> 
#> D4: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.184   0.064   0.067   0.308
#>    B2     0.050   0.068  -0.078   0.187
#>    B3     0.024   0.072  -0.102   0.159
#>    B4     0.062   0.071  -0.059   0.210
#>    B5    -0.006   0.063  -0.134   0.109
#>    C1    -0.049   0.054  -0.157   0.051
#>    C2     0.002   0.059  -0.137   0.109
#>    D1     0.006   0.052  -0.102   0.104
#>    D2     0.188   0.055   0.087   0.297
#>    D3     0.422   0.056   0.306   0.520
#>    D5     0.038   0.067  -0.093   0.153
#>    D6     0.063   0.072  -0.097   0.203
#>    D7    -0.047   0.064  -0.176   0.076
#>    E1     0.151   0.063   0.013   0.262
#>    E2    -0.030   0.053  -0.130   0.072
#>    E3     0.020   0.053  -0.088   0.128
#>    E4     0.059   0.062  -0.052   0.172
#>    E5     0.004   0.064  -0.118   0.125
#>    E6    -0.043   0.056  -0.143   0.059
#> 
#> D5: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.093   0.068  -0.020   0.231
#>    B2     0.089   0.078  -0.069   0.241
#>    B3     0.130   0.077  -0.025   0.280
#>    B4    -0.124   0.078  -0.264   0.024
#>    B5    -0.026   0.074  -0.170   0.121
#>    C1     0.001   0.060  -0.105   0.122
#>    C2     0.179   0.059   0.051   0.286
#>    D1    -0.015   0.053  -0.112   0.096
#>    D2     0.196   0.067   0.060   0.330
#>    D3    -0.135   0.072  -0.280   0.002
#>    D4     0.049   0.084  -0.119   0.192
#>    D6     0.120   0.071  -0.029   0.244
#>    D7     0.069   0.069  -0.061   0.198
#>    E1     0.045   0.072  -0.108   0.165
#>    E2    -0.008   0.062  -0.125   0.122
#>    E3     0.000   0.061  -0.121   0.119
#>    E4     0.008   0.069  -0.117   0.159
#>    E5     0.291   0.068   0.156   0.410
#>    E6     0.013   0.059  -0.106   0.118
#> 
#> D6: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.054   0.067  -0.064   0.185
#>    B2     0.051   0.073  -0.094   0.199
#>    B3    -0.062   0.066  -0.181   0.061
#>    B4    -0.087   0.073  -0.241   0.047
#>    B5    -0.051   0.071  -0.195   0.084
#>    C1     0.069   0.057  -0.037   0.180
#>    C2    -0.009   0.069  -0.143   0.126
#>    D1     0.002   0.052  -0.090   0.098
#>    D2     0.058   0.063  -0.071   0.194
#>    D3    -0.002   0.068  -0.129   0.137
#>    D4     0.069   0.078  -0.102   0.242
#>    D5     0.106   0.064  -0.027   0.221
#>    D7     0.426   0.055   0.317   0.531
#>    E1     0.007   0.067  -0.124   0.137
#>    E2     0.046   0.053  -0.051   0.143
#>    E3     0.047   0.068  -0.084   0.188
#>    E4     0.138   0.066  -0.014   0.255
#>    E5     0.116   0.062  -0.002   0.231
#>    E6     0.092   0.057  -0.026   0.200
#> 
#> D7: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.034   0.076  -0.118   0.184
#>    B2    -0.104   0.077  -0.269   0.033
#>    B3     0.116   0.076  -0.029   0.273
#>    B4    -0.120   0.077  -0.254   0.018
#>    B5    -0.022   0.081  -0.177   0.123
#>    C1     0.023   0.060  -0.084   0.141
#>    C2     0.082   0.066  -0.046   0.212
#>    D1    -0.042   0.058  -0.144   0.064
#>    D2     0.149   0.067   0.021   0.277
#>    D3    -0.048   0.069  -0.181   0.081
#>    D4    -0.058   0.080  -0.219   0.101
#>    D5     0.071   0.069  -0.058   0.209
#>    D6     0.487   0.065   0.362   0.592
#>    E1     0.129   0.078  -0.010   0.277
#>    E2     0.061   0.062  -0.050   0.191
#>    E3     0.009   0.060  -0.113   0.125
#>    E4    -0.104   0.067  -0.220   0.025
#>    E5     0.057   0.076  -0.089   0.202
#>    E6     0.050   0.063  -0.070   0.176
#> 
#> E1: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.057   0.078  -0.097   0.212
#>    B2    -0.087   0.085  -0.261   0.069
#>    B3     0.013   0.084  -0.132   0.194
#>    B4    -0.004   0.084  -0.168   0.168
#>    B5     0.207   0.079   0.066   0.366
#>    C1    -0.034   0.066  -0.165   0.092
#>    C2     0.114   0.063  -0.008   0.233
#>    D1     0.026   0.057  -0.098   0.120
#>    D2     0.114   0.066  -0.006   0.245
#>    D3    -0.080   0.073  -0.237   0.072
#>    D4     0.205   0.087   0.018   0.376
#>    D5     0.050   0.081  -0.121   0.193
#>    D6     0.010   0.084  -0.157   0.161
#>    D7     0.139   0.081  -0.011   0.296
#>    E2     0.231   0.064   0.108   0.342
#>    E3    -0.004   0.063  -0.127   0.130
#>    E4     0.033   0.067  -0.096   0.176
#>    E5     0.011   0.069  -0.134   0.131
#>    E6     0.023   0.068  -0.102   0.159
#> 
#> E2: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.028   0.086  -0.196   0.142
#>    B2     0.219   0.096   0.033   0.425
#>    B3     0.000   0.091  -0.186   0.162
#>    B4    -0.064   0.089  -0.237   0.092
#>    B5    -0.056   0.085  -0.222   0.110
#>    C1     0.002   0.074  -0.154   0.123
#>    C2     0.048   0.078  -0.096   0.185
#>    D1     0.104   0.065  -0.016   0.233
#>    D2    -0.043   0.079  -0.205   0.098
#>    D3     0.143   0.082  -0.012   0.293
#>    D4    -0.052   0.092  -0.225   0.115
#>    D5    -0.010   0.086  -0.173   0.165
#>    D6     0.073   0.083  -0.077   0.231
#>    D7     0.084   0.087  -0.074   0.267
#>    E1     0.292   0.082   0.128   0.443
#>    E3     0.053   0.074  -0.095   0.195
#>    E4     0.146   0.086  -0.025   0.328
#>    E5    -0.057   0.083  -0.218   0.096
#>    E6    -0.070   0.071  -0.203   0.068
#> 
#> E3: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.044   0.090  -0.207   0.124
#>    B2    -0.020   0.080  -0.179   0.129
#>    B3     0.142   0.090  -0.026   0.305
#>    B4    -0.004   0.092  -0.193   0.161
#>    B5    -0.053   0.089  -0.224   0.132
#>    C1     0.108   0.068  -0.028   0.229
#>    C2     0.024   0.075  -0.118   0.159
#>    D1     0.089   0.066  -0.052   0.212
#>    D2     0.117   0.082  -0.043   0.285
#>    D3    -0.054   0.077  -0.213   0.096
#>    D4     0.033   0.090  -0.152   0.208
#>    D5     0.001   0.084  -0.165   0.161
#>    D6     0.072   0.102  -0.118   0.262
#>    D7     0.012   0.080  -0.143   0.162
#>    E1    -0.006   0.078  -0.150   0.165
#>    E2     0.051   0.073  -0.096   0.194
#>    E4     0.357   0.073   0.217   0.505
#>    E5    -0.041   0.084  -0.193   0.119
#>    E6     0.048   0.074  -0.101   0.186
#> 
#> E4: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1    -0.163   0.078  -0.332  -0.026
#>    B2     0.114   0.075  -0.045   0.248
#>    B3     0.064   0.087  -0.103   0.220
#>    B4     0.042   0.091  -0.140   0.221
#>    B5     0.195   0.076   0.054   0.335
#>    C1    -0.033   0.062  -0.162   0.082
#>    C2     0.042   0.068  -0.088   0.179
#>    D1    -0.021   0.062  -0.132   0.101
#>    D2    -0.015   0.070  -0.141   0.129
#>    D3    -0.130   0.070  -0.251   0.015
#>    D4     0.079   0.083  -0.069   0.245
#>    D5     0.007   0.073  -0.127   0.152
#>    D6     0.166   0.080  -0.013   0.303
#>    D7    -0.109   0.070  -0.236   0.030
#>    E1     0.032   0.066  -0.088   0.170
#>    E2     0.112   0.066  -0.017   0.251
#>    E3     0.278   0.058   0.164   0.386
#>    E5     0.152   0.072   0.019   0.294
#>    E6     0.102   0.068  -0.012   0.240
#> 
#> E5: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.037   0.075  -0.104   0.186
#>    B2    -0.079   0.082  -0.255   0.063
#>    B3     0.074   0.083  -0.090   0.233
#>    B4    -0.064   0.081  -0.212   0.086
#>    B5     0.048   0.086  -0.149   0.219
#>    C1    -0.010   0.062  -0.133   0.121
#>    C2    -0.152   0.065  -0.282  -0.029
#>    D1     0.116   0.059  -0.001   0.225
#>    D2     0.044   0.068  -0.090   0.165
#>    D3     0.107   0.071  -0.050   0.237
#>    D4     0.004   0.087  -0.172   0.167
#>    D5     0.317   0.076   0.175   0.458
#>    D6     0.145   0.078  -0.003   0.303
#>    D7     0.060   0.081  -0.095   0.216
#>    E1     0.011   0.069  -0.136   0.133
#>    E2    -0.043   0.063  -0.169   0.080
#>    E3    -0.033   0.068  -0.159   0.098
#>    E4     0.157   0.073   0.020   0.293
#>    E6     0.175   0.067   0.036   0.299
#> 
#> E6: 
#>  Node Post.mean Post.sd Cred.lb Cred.ub
#>    B1     0.062   0.091  -0.111   0.255
#>    B2     0.173   0.083   0.029   0.347
#>    B3     0.064   0.086  -0.113   0.220
#>    B4    -0.003   0.088  -0.204   0.149
#>    B5     0.090   0.083  -0.072   0.246
#>    C1    -0.019   0.069  -0.152   0.118
#>    C2    -0.072   0.076  -0.241   0.064
#>    D1    -0.029   0.058  -0.120   0.091
#>    D2    -0.023   0.081  -0.183   0.126
#>    D3     0.041   0.072  -0.097   0.168
#>    D4    -0.071   0.094  -0.244   0.097
#>    D5     0.016   0.079  -0.136   0.155
#>    D6     0.138   0.083  -0.041   0.294
#>    D7     0.065   0.082  -0.088   0.219
#>    E1     0.029   0.081  -0.117   0.191
#>    E2    -0.066   0.066  -0.185   0.069
#>    E3     0.046   0.071  -0.100   0.172
#>    E4     0.126   0.083  -0.018   0.294
#>    E5     0.211   0.081   0.040   0.344
#> 


# ordinal

# fit model (note + 1, due to zeros)
fit <- explore(Y + 1, type = "ordinal",
               iter = 250)

#> BGGM: Posterior Sampling 

#> BGGM: Prior Sampling 

#> BGGM: Finished


# summarize the partial correlations
reg <- coef(fit)

#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |==========                                                            |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |========================                                              |  35%  |                                                                              |============================                                          |  40%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================                                |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |====================================================                  |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |==================================================================    |  95%  |                                                                              |======================================================================| 100%


# summary
summ <- summary(reg)

#########################
### example 2: binary ###
#########################
# data
Y <- women_math

# fit model
fit <- explore(Y, type = "binary",
               iter = 250)

#> BGGM: Posterior Sampling 

#> BGGM: Prior Sampling 

#> BGGM: Finished


# summarize the partial correlations
reg <- coef(fit)

#>   |                                                                              |                                                                      |   0%  |                                                                              |============                                                          |  17%  |                                                                              |=======================                                               |  33%  |                                                                              |===================================                                   |  50%  |                                                                              |===============================================                       |  67%  |                                                                              |==========================================================            |  83%  |                                                                              |======================================================================| 100%


# summary
summ <- summary(reg)

# }