GGM Compare: Confirmatory Hypothesis Testing

Confirmatory hypothesis testing for comparing GGMs. Hypotheses are expressed as equality and/or ineqaulity contraints on the partial correlations of interest. Here the focus is not on determining the graph (see explore) but testing specific hypotheses related to the conditional (in)dependence structure. These methods were introduced in Williams2019_bf;textualBGGM and in williams2020comparing;textualBGGM

Usage

ggm_compare_confirm(
  ...,
  hypothesis,
  formula = NULL,
  type = "continuous",
  mixed_type = NULL,
  prior_sd = 0.5,
  iter = 25000,
  impute = TRUE,
  progress = TRUE,
  seed = NULL
)

Arguments

...

At least two matrices (or data frame) of dimensions n (observations) by p (nodes).

hypothesis

Character string. The hypothesis (or hypotheses) to be tested. See notes for futher details.

formula

an object of class formula. This allows for including control variables in the model (i.e., ~ gender).

type

Character string. Which type of data for Y ? The options include continuous, binary, ordinal, or mixed. Note that mixed can be used for data with only ordinal variables. See the note for further details.

mixed_type

numeric vector. An indicator of length p for which varibles should be treated as ranks. (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables as ranks when type = "mixed" and NULL otherwise. See note for further details.

prior_sd

Numeric. The scale of the prior distribution (centered at zero), in reference to a beta distribtuion (defaults to 0.5).

iter

Number of iterations (posterior samples; defaults to 25,000).

impute

Logicial. Should the missing values (NA) be imputed during model fitting (defaults to TRUE) ?

progress

Logical. Should a progress bar be included (defaults to TRUE) ?

seed

An integer for the random seed.

Value

The returned object of class confirm contains a lot of information that is used for printing and plotting the results. For users of BGGM, the following are the useful objects:

out_hyp_prob Posterior hypothesis probabilities.

info An object of class BF from the R package BFpack mulder2019bfpackBGGM

Details

The hypotheses can be written either with the respective column names or numbers. For example, g1_1--2 denotes the relation between the variables in column 1 and 2 for group 1. The g1_ is required and the only difference from confirm (one group). Note that these must correspond to the upper triangular elements of the correlation matrix. This is accomplished by ensuring that the first number is smaller than the second number. This also applies when using column names (i.e,, in reference to the column number).

One Hypothesis:

To test whether a relation in larger in one group, while both are expected to be positive, this can be written as

hyp <- c(g1_1--2 > g2_1--2 > 0)

This is then compared to the complement.

More Than One Hypothesis:

The above hypothesis can also be compared to, say, a null model by using ";" to seperate the hypotheses, for example,

hyp <- c(g1_1--2 > g2_1--2 > 0; g1_1--2 = g2_1--2 = 0).

Any number of hypotheses can be compared this way.

Using "&"

It is also possible to include &. This allows for testing one constraint and another contraint as one hypothesis.

hyp <- c("g1_A1--A2 > g2_A1--A2 & g1_A1--A3 = g2_A1--A3")

Of course, it is then possible to include additional hypotheses by separating them with ";".

Testing Sums

It might also be interesting to test the sum of partial correlations. For example, that the sum of specific relations in one group is larger than the sum in another group.

hyp <- c("g1_A1--A2 + g1_A1--A3 > g2_A1--A2 + g2_A1--A3; g1_A1--A2 + g1_A1--A3 = g2_A1--A2 + g2_A1--A3")

Potential Delays:

There is a chance for a potentially long delay from the time the progress bar finishes to when the function is done running. This occurs when the hypotheses require further sampling to be tested, for example, when grouping relations c("(g1_A1--A2, g2_A2--A3) > (g2_A1--A2, g2_A2--A3)". This is not an error.

Controlling for Variables:

When controlling for variables, it is assumed that Y includes only the nodes in the GGM and the control variables. Internally, only the predictors that are included in formula are removed from Y. This is not behavior of, say, lm, but was adopted to ensure users do not have to write out each variable that should be included in the GGM. An example is provided below.

Mixed Type:

The term "mixed" is somewhat of a misnomer, because the method can be used for data including only continuous or only discrete variables hoff2007extendingBGGM. This is based on the ranked likelihood which requires sampling the ranks for each variable (i.e., the data is not merely transformed to ranks). This is computationally expensive when there are many levels. For example, with continuous data, there are as many ranks as data points!

The option mixed_type allows the user to determine which variable should be treated as ranks and the "emprical" distribution is used otherwise. This is accomplished by specifying an indicator vector of length p. A one indicates to use the ranks, whereas a zero indicates to "ignore" that variable. By default all integer variables are handled as ranks.

Dealing with Errors:

An error is most likely to arise when type = "ordinal". The are two common errors (although still rare):

The first is due to sampling the thresholds, especially when the data is heavily skewed. This can result in an ill-defined matrix. If this occurs, we recommend to first try decreasing prior_sd (i.e., a more informative prior). If that does not work, then change the data type to type = mixed which then estimates a copula GGM (this method can be used for data containing only ordinal variable). This should work without a problem.

The second is due to how the ordinal data are categorized. For example, if the error states that the index is out of bounds, this indicates that the first category is a zero. This is not allowed, as the first category must be one. This is addressed by adding one (e.g., Y + 1) to the data matrix.

Imputing Missing Values:

Missing values are imputed with the approach described in hoff2009first;textualBGGM. The basic idea is to impute the missing values with the respective posterior pedictive distribution, given the observed data, as the model is being estimated. Note that the default is TRUE, but this ignored when there are no missing values. If set to FALSE, and there are missing values, list-wise deletion is performed with na.omit.

Note

"Default" Prior:

In Bayesian statistics, a default Bayes factor needs to have several properties. I refer interested users to @section 2.2 in @dablander2020default;textualBGGM. In Williams2019_bf;textualBGGM, some of these propteries were investigated (e.g., model selection consistency). That said, we would not consider this a "default" or "automatic" Bayes factor and thus we encourage users to perform sensitivity analyses by varying the scale of the prior distribution (prior_sd).

Furthermore, it is important to note there is no "correct" prior and, also, there is no need to entertain the possibility of a "true" model. Rather, the Bayes factor can be interpreted as which hypothesis best (relative to each other) predicts the observed data @Section 3.2 in @Kass1995BGGM.

Interpretation of Conditional (In)dependence Models for Latent Data:

See BGGM-package for details about interpreting GGMs based on latent data (i.e, all data types besides "continuous")

References

Examples

# \donttest{
# note: iter = 250 for demonstrative purposes

# data
Y <- bfi

###############################
#### example 1: continuous ####
###############################

# males
Ymale   <- subset(Y, gender == 1,
                  select = -c(education,
                              gender))[,1:5]

# females
Yfemale <- subset(Y, gender == 2,
                     select = -c(education,
                                 gender))[,1:5]

 # exhaustive
 hypothesis <- c("g1_A1--A2 >  g2_A1--A2;
                  g1_A1--A2 <  g2_A1--A2;
                  g1_A1--A2 =  g2_A1--A2")

# test hyp
test <- ggm_compare_confirm(Ymale,  Yfemale,
                            hypothesis = hypothesis,
                            iter = 250,
                            progress = FALSE)

# print (evidence not strong)
test
#> BGGM: Bayesian Gaussian Graphical Models 
#> Type: continuous 
#> --- 
#> Posterior Samples: 250 
#>   Group 1: 896 
#>   Group 2: 1813 
#> Variables (p): 5 
#> Relations: 10 
#> Delta: 3 
#> --- 
#> Call:
#> ggm_compare_confirm(Ymale, Yfemale, hypothesis = hypothesis, 
#>     iter = 250, progress = FALSE)
#> --- 
#> Hypotheses: 
#> 
#> H1: g1_A1--A2>g2_A1--A2
#> H2: g1_A1--A2<g2_A1--A2
#> H3: g1_A1--A2=g2_A1--A2
#> --- 
#> Posterior prob: 
#> 
#> p(H1|data) = 0.136
#> p(H2|data) = 0.016
#> p(H3|data) = 0.848
#> --- 
#> Bayes factor matrix: 
#>       H1     H2    H3
#> H1 1.000  8.528 0.161
#> H2 0.117  1.000 0.019
#> H3 6.215 53.004 1.000
#> --- 
#> note: equal hypothesis prior probabilities

#########################################
#### example 2: sensitivity to prior ####
#########################################
# continued from example 1

# decrease prior SD
test <- ggm_compare_confirm(Ymale,
                            Yfemale,
                            prior_sd = 0.1,
                            hypothesis = hypothesis,
                            iter = 250,
                            progress = FALSE)

# print
test
#> BGGM: Bayesian Gaussian Graphical Models 
#> Type: continuous 
#> --- 
#> Posterior Samples: 250 
#>   Group 1: 896 
#>   Group 2: 1813 
#> Variables (p): 5 
#> Relations: 10 
#> Delta: 99 
#> --- 
#> Call:
#> ggm_compare_confirm(Ymale, Yfemale, hypothesis = hypothesis, 
#>     prior_sd = 0.1, iter = 250, progress = FALSE)
#> --- 
#> Hypotheses: 
#> 
#> H1: g1_A1--A2>g2_A1--A2
#> H2: g1_A1--A2<g2_A1--A2
#> H3: g1_A1--A2=g2_A1--A2
#> --- 
#> Posterior prob: 
#> 
#> p(H1|data) = 0.434
#> p(H2|data) = 0.046
#> p(H3|data) = 0.52
#> --- 
#> Bayes factor matrix: 
#>       H1     H2    H3
#> H1 1.000  9.372 0.835
#> H2 0.107  1.000 0.089
#> H3 1.197 11.219 1.000
#> --- 
#> note: equal hypothesis prior probabilities

# indecrease prior SD
test <- ggm_compare_confirm(Ymale,
                            Yfemale,
                            prior_sd = 0.28,
                            hypothesis = hypothesis,
                            iter = 250,
                            progress = FALSE)

# print
test
#> BGGM: Bayesian Gaussian Graphical Models 
#> Type: continuous 
#> --- 
#> Posterior Samples: 250 
#>   Group 1: 896 
#>   Group 2: 1813 
#> Variables (p): 5 
#> Relations: 10 
#> Delta: 11.7551 
#> --- 
#> Call:
#> ggm_compare_confirm(Ymale, Yfemale, hypothesis = hypothesis, 
#>     prior_sd = 0.28, iter = 250, progress = FALSE)
#> --- 
#> Hypotheses: 
#> 
#> H1: g1_A1--A2>g2_A1--A2
#> H2: g1_A1--A2<g2_A1--A2
#> H3: g1_A1--A2=g2_A1--A2
#> --- 
#> Posterior prob: 
#> 
#> p(H1|data) = 0.26
#> p(H2|data) = 0.03
#> p(H3|data) = 0.71
#> --- 
#> Bayes factor matrix: 
#>       H1     H2    H3
#> H1 1.000  8.730 0.367
#> H2 0.115  1.000 0.042
#> H3 2.728 23.818 1.000
#> --- 
#> note: equal hypothesis prior probabilities

################################
#### example 3: mixed data #####
################################

hypothesis <- c("g1_A1--A2 >  g2_A1--A2;
                 g1_A1--A2 <  g2_A1--A2;
                 g1_A1--A2 =  g2_A1--A2")

# test (1000 for example)
test <- ggm_compare_confirm(Ymale,
                            Yfemale,
                            type = "mixed",
                            hypothesis = hypothesis,
                            iter = 250,
                            progress = FALSE)
#> Warning: imputation during model fitting is
#> currently only implemented for 'continuous' data.

# print
test
#> BGGM: Bayesian Gaussian Graphical Models 
#> Type: mixed 
#> --- 
#> Posterior Samples: 250 
#>   Group 1: 896 
#>   Group 2: 1813 
#> Variables (p): 5 
#> Relations: 10 
#> Delta: 3 
#> --- 
#> Call:
#> ggm_compare_confirm(Ymale, Yfemale, hypothesis = hypothesis, 
#>     type = "mixed", iter = 250, progress = FALSE)
#> --- 
#> Hypotheses: 
#> 
#> H1: g1_A1--A2>g2_A1--A2
#> H2: g1_A1--A2<g2_A1--A2
#> H3: g1_A1--A2=g2_A1--A2
#> --- 
#> Posterior prob: 
#> 
#> p(H1|data) = 0.028
#> p(H2|data) = 0.091
#> p(H3|data) = 0.882
#> --- 
#> Bayes factor matrix: 
#>        H1    H2    H3
#> H1  1.000 0.307 0.032
#> H2  3.261 1.000 0.103
#> H3 31.734 9.732 1.000
#> --- 
#> note: equal hypothesis prior probabilities

##############################
##### example 4: control #####
##############################
# control for education

# data
Y <- bfi

# males
Ymale   <- subset(Y, gender == 1,
                  select = -c(gender))[,c(1:5, 26)]

# females
Yfemale <- subset(Y, gender == 2,
                  select = -c(gender))[,c(1:5, 26)]

# test
test <- ggm_compare_confirm(Ymale,
                             Yfemale,
                             formula = ~ education,
                             hypothesis = hypothesis,
                             iter = 250,
                             progress = FALSE)
# print
test
#> BGGM: Bayesian Gaussian Graphical Models 
#> Type: continuous 
#> --- 
#> Posterior Samples: 250 
#>   Group 1: 817 
#>   Group 2: 1676 
#> Variables (p): 5 
#> Relations: 10 
#> Delta: 3 
#> --- 
#> Call:
#> ggm_compare_confirm(Ymale, Yfemale, hypothesis = hypothesis, 
#>     formula = ~education, iter = 250, progress = FALSE)
#> --- 
#> Hypotheses: 
#> 
#> H1: g1_A1--A2>g2_A1--A2
#> H2: g1_A1--A2<g2_A1--A2
#> H3: g1_A1--A2=g2_A1--A2
#> --- 
#> Posterior prob: 
#> 
#> p(H1|data) = 0.025
#> p(H2|data) = 0.07
#> p(H3|data) = 0.905
#> --- 
#> Bayes factor matrix: 
#>        H1     H2    H3
#> H1  1.000  0.360 0.028
#> H2  2.779  1.000 0.077
#> H3 36.044 12.972 1.000
#> --- 
#> note: equal hypothesis prior probabilities

#####################################
##### example 5: many relations #####
#####################################

# data
Y <- bfi

hypothesis <- c("g1_A1--A2 > g2_A1--A2 & g1_A1--A3 = g2_A1--A3;
                 g1_A1--A2 = g2_A1--A2 & g1_A1--A3 = g2_A1--A3;
                 g1_A1--A2 = g2_A1--A2 = g1_A1--A3 = g2_A1--A3")

Ymale   <- subset(Y, gender == 1,
                  select = -c(education,
                              gender))[,1:5]

# females
Yfemale <- subset(Y, gender == 2,
                     select = -c(education,
                                 gender))[,1:5]

test <- ggm_compare_confirm(Ymale,
                            Yfemale,
                             hypothesis = hypothesis,
                             iter = 250,
                             progress = FALSE)

# print
test
#> BGGM: Bayesian Gaussian Graphical Models 
#> Type: continuous 
#> --- 
#> Posterior Samples: 250 
#>   Group 1: 896 
#>   Group 2: 1813 
#> Variables (p): 5 
#> Relations: 10 
#> Delta: 3 
#> --- 
#> Call:
#> ggm_compare_confirm(Ymale, Yfemale, hypothesis = hypothesis, 
#>     iter = 250, progress = FALSE)
#> --- 
#> Hypotheses: 
#> 
#> H1: g1_A1--A2>g2_A1--A2&g1_A1--A3=g2_A1--A3
#> H2: g1_A1--A2=g2_A1--A2&g1_A1--A3=g2_A1--A3
#> H3: g1_A1--A2=g2_A1--A2=g1_A1--A3=g2_A1--A3
#> H4: complement
#> --- 
#> Posterior prob: 
#> 
#> p(H1|data) = 0.118
#> p(H2|data) = 0.873
#> p(H3|data) = 0.006
#> p(H4|data) = 0.003
#> --- 
#> Bayes factor matrix: 
#>       H1    H2      H3      H4
#> H1 1.000 0.135  19.490  36.820
#> H2 7.414 1.000 144.499 272.984
#> H3 0.051 0.007   1.000   1.889
#> H4 0.027 0.004   0.529   1.000
#> --- 
#> note: equal hypothesis prior probabilities
# }