Extract the Partial Correlation Matrix

Extract the partial correlation matrix (posterior mean) from estimate, explore, ggm_compare_estimate, and ggm_compare_explore objects. It is also possible to extract the partial correlation differences for ggm_compare_estimate and ggm_compare_explore objects.

Usage

pcor_mat(object, difference = FALSE, ...)

Arguments

object

A model estimated with BGGM. All classes are supported, assuming there is matrix to be extracted.

difference

Logical. Should the difference be returned (defaults to FALSE) ? Note that this assumes there is a difference (e.g., an object of class ggm_compare_estimate) and ignored otherwise.

...

Currently ignored.

Value

The estimated partial correlation matrix.

Examples

# \donttest{
# note: iter = 250 for demonstrative purposes

# data
Y <- ptsd[,1:5] + 1

# ordinal
fit <- estimate(Y, type = "ordinal",
                iter = 250,
                progress = FALSE)

pcor_mat(fit)
#>       B1     B2    B3     B4    B5
#> B1 0.000  0.263 0.035  0.348 0.119
#> B2 0.263  0.000 0.538 -0.103 0.120
#> B3 0.035  0.538 0.000  0.273 0.184
#> B4 0.348 -0.103 0.273  0.000 0.366
#> B5 0.119  0.120 0.184  0.366 0.000
# }