Summarary Method for Multivariate or Univarate Regression

Summarary Method for Multivariate or Univarate Regression

Usage

regression_summary(object, cred = 0.95, ...)

Arguments

object

An object of class estimate

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored

Value

A list of length p including the summaries for each regression.

Examples

# \donttest{
# note: iter = 250 for demonstrative purposes

# data
Y <- bfi

Y <- subset(Y, select = c("A1", "A2", 
                          "gender", "education"))

fit_mv_ordinal <- estimate(Y, formula = ~ gender + as.factor(education),
                           type = "continuous",
                           iter = 250,
                           progress = TRUE)
#> BGGM: Posterior Sampling 
#> BGGM: Finished

regression_summary(fit_mv_ordinal)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Type: continuous 
#> Formula: ~ gender + as.factor(education) 
#> --- 
#> Coefficients: 
#>  
#> A1 
#>                       Post.mean Post.sd Cred.lb Cred.ub
#> (Intercept)               1.039   0.137   0.779   1.327
#> gender                   -0.504   0.060  -0.616  -0.392
#> as.factor(education)2     0.129   0.126  -0.110   0.371
#> as.factor(education)3    -0.126   0.106  -0.316   0.084
#> as.factor(education)4    -0.424   0.114  -0.631  -0.197
#> as.factor(education)5    -0.541   0.114  -0.745  -0.312
#> --- 
#> A2 
#>                       Post.mean Post.sd Cred.lb Cred.ub
#> (Intercept)              -0.865   0.108  -1.080  -0.651
#> gender                    0.484   0.051   0.380   0.569
#> as.factor(education)2    -0.030   0.100  -0.224   0.145
#> as.factor(education)3     0.108   0.082  -0.063   0.258
#> as.factor(education)4    -0.043   0.091  -0.228   0.131
#> as.factor(education)5     0.082   0.091  -0.095   0.263
#> --- 
#> Residual Correlation Matrix: 
#>        A1     A2
#> A1  1.000 -0.314
#> A2 -0.314  1.000
#> --- 
# }