Graph Selection for estimate Objects

Provides the selected graph based on credible intervals for the partial correlations that did not contain zero Williams2019BGGM.

Usage

# S3 method for class 'estimate'
select(object, cred = 0.95, alternative = "two.sided", ...)

Arguments

object

An object of class estimate.default.

cred

Numeric. The credible interval width for selecting the graph (defaults to 0.95; must be between 0 and 1).

alternative

A character string specifying the alternative hypothesis. It must be one of "two.sided" (default), "greater" or "less". See note for futher details.

...

Currently ignored.

Value

The returned object of class select.estimate contains a lot of information that is used for printing and plotting the results. For users of BGGM, the following are the useful objects:

pcor_adj Selected partial correlation matrix (weighted adjacency).

adj Adjacency matrix for the selected edges

object An object of class estimate (the fitted model).

Details

This package was built for the social-behavioral sciences in particular. In these applications, there is strong theory that expects all effects to be positive. This is known as a "positive manifold" and this notion has a rich tradition in psychometrics. Hence, this can be incorporated into the graph with alternative = "greater". This results in the estimated structure including only positive edges.

References

See also

Examples

# \donttest{
# note: iter = 250 for demonstrative purposes

# data
Y <- bfi[,1:10]

# estimate
fit <- estimate(Y, iter = 250,
                progress = FALSE)

# select edge set
E <- select(fit)

# }