Summary method for estimate.default objects

Summarize the posterior distribution of each partial correlation with the posterior mean and standard deviation.

Usage

# S3 method for class 'estimate'
summary(object, col_names = TRUE, cred = 0.95, ...)

Arguments

object

An object of class estimate

col_names

Logical. Should the summary include the column names (default is TRUE)? Setting to FALSE includes the column numbers (e.g., 1--2).

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored.

Value

A dataframe containing the summarized posterior distributions.

See also

Examples

# \donttest{
# data
Y <- ptsd[,1:5]

fit <- estimate(Y, iter = 250,
                progress = FALSE)

summary(fit)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Type: continuous 
#> Analytic: FALSE 
#> Formula:  
#> Posterior Samples: 250 
#> Observations (n):
#> Nodes (p): 5 
#> Relations: 10 
#> --- 
#> Call: 
#> estimate(Y = Y, iter = 250, progress = FALSE)
#> --- 
#> Estimates:
#>  Relation Post.mean Post.sd Cred.lb Cred.ub
#>    B1--B2     0.230   0.063   0.098   0.342
#>    B1--B3     0.063   0.069  -0.065   0.195
#>    B2--B3     0.489   0.050   0.391   0.581
#>    B1--B4     0.332   0.061   0.206   0.439
#>    B2--B4    -0.031   0.073  -0.161   0.114
#>    B3--B4     0.215   0.066   0.091   0.351
#>    B1--B5     0.149   0.069   0.008   0.284
#>    B2--B5     0.107   0.068  -0.022   0.229
#>    B3--B5     0.188   0.067   0.055   0.314
#>    B4--B5     0.333   0.058   0.225   0.446
#> --- 

# }