Summary method for ggm_compare_estimate objects

Summarize the posterior distribution of each partial correlation difference with the posterior mean and standard deviation.

# S3 method for ggm_compare_estimate
summary(object, col_names = TRUE, cred = 0.95, ...)

Arguments

object
An object of class ggm_compare_estimate.

col_names
Logical. Should the summary include the column names (default is TRUE)? Setting to FALSE includes the column numbers (e.g., 1--2).

cred
Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...
Currently ignored.

Value

A list containing the summarized posterior distributions.

See also

Examples

# \donttest{
# note: iter = 250 for demonstrative purposes
# data
Y <- bfi

# males and females
Ymale <- subset(Y, gender == 1,
                select = -c(gender,
                            education))[,1:5]

Yfemale <- subset(Y, gender == 2,
                  select = -c(gender,
                              education))[,1:5]

# fit model
fit <- ggm_compare_estimate(Ymale,  Yfemale,
                            type = "ordinal",
                            iter = 250,
                            prior_sd = 0.25,
                            progress = FALSE)

summary(fit)

#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Type: ordinal 
#> Analytic: FALSE 
#> Formula:  
#> Posterior Samples: 250 
#> Observations (n):
#>   Group 1: 896 
#>   Group 2: 1813 
#> Nodes (p): 5 
#> Relations: 10 
#> --- 
#> Call: 
#> ggm_compare_estimate(Ymale, Yfemale, type = "ordinal", prior_sd = 0.25, 
#>     iter = 250, progress = FALSE)
#> --- 
#> Estimates:
#> 
#>  
#>  Relation Post.mean Post.sd Cred.lb Cred.ub
#>  A1--A2    0.046    0.041   -0.030  0.138  
#>  A1--A3   -0.005    0.049   -0.098  0.086  
#>  A2--A3   -0.064    0.044   -0.147  0.025  
#>  A1--A4    0.007    0.044   -0.084  0.091  
#>  A2--A4   -0.007    0.045   -0.093  0.083  
#>  A3--A4    0.028    0.050   -0.068  0.125  
#>  A1--A5    0.031    0.053   -0.068  0.127  
#>  A2--A5    0.099    0.047    0.008  0.180  
#>  A3--A5    0.071    0.040    0.002  0.153  
#>  A4--A5   -0.022    0.043   -0.101  0.062  
#> --- 

# }