Data: Toronto Alexithymia Scale (TAS)

A dataset containing items from the Toronto Alexithymia Scale (TAS). There are 20 variables and 1925 observations

data("tas")

Format

A data frame with 20 variables and 1925 observations (5 point Likert scale)

Details

1 I am often confused about what emotion I am feeling

2 It is difficult for me to find the right words for my feelings

3 I have physical sensations that even doctors don’t understand

4 I am able to describe my feelings easily

5 I prefer to analyze problems rather than just describe them

6 When I am upset, I don’t know if I am sad, frightened, or angry

7 I am often puzzled by sensations in my body

8 I prefer just to let things happen rather than to understand why they turned out that way

9 I have feelings that I can’t quite identify

10 Being in touch with emotions is essential

11 I find it hard to describe how I feel about people

12 People tell me to describe my feelings more

13 I don’t know what’s going on inside me

14 I often don’t know why I am angry

15 I prefer talking to people about their daily activities rather than their feelings

16 I prefer to watch “light” entertainment shows rather than psychological dramas

17 It is difficult for me to reveal my innermost feelings, even to close friends

18 I can feel close to someone, even in moments of silence

19 I find examination of my feelings useful in solving personal problems

20 Looking for hidden meanings in movies or plays distracts from their enjoyment

gender "M" (male) or "F" (female)

Note

There are three domains

Difficulty identifying feelings: items 1, 3, 6, 7, 9, 13, 14

Difficulty describing feelings: items 2, 4, 11, 12, 17

Externally oriented thinking: items 10, 15, 16, 18, 19

References

Briganti, G., & Linkowski, P. (2019). Network approach to items and domains from the Toronto Alexithymia Scale. Psychological reports.

Examples

data("tas")

# labels
tas_labels <- BGGM:::tas_labels