Data: Interpersonal Reactivity Index (IRI)

A dataset containing items from the Interpersonal Reactivity Index (IRI; an empathy measure). There are 28 variables and 1973 observations

data("iri")

Format

A data frame with 28 variables and 1973 observations (5 point Likert scale)

Details

1 I daydream and fantasize, with some regularity, about things that might happen to me.

2 I often have tender, concerned feelings for people less fortunate than me.

3 I sometimes find it difficult to see things from the "other guy's" point of view.

4 Sometimes I don't feel very sorry for other people when they are having problems.

5 I really get involved with the feelings of the characters in a novel.

6 In emergency situations, I feel apprehensive and ill-at-ease.

7 I am usually objective when I watch a movie or play, and I don't often get completely caught up in it.

8 I try to look at everybody's side of a disagreement before I make a decision.

9 When I see someone being taken advantage of, I feel kind of protective towards them.

10 I sometimes feel helpless when I am in the middle of a very emotional situation.

11 I sometimes try to understand my friends better by imagining how things look from their perspective

12 Becoming extremely involved in a good book or movie is somewhat rare for me.

13 When I see someone get hurt, I tend to remain calm.

14 Other people's misfortunes do not usually disturb me a great deal.

15 If I'm sure I'm right about something, I don't waste much time listening to other people's arguments.

16 After seeing a play or movie, I have felt as though I were one of the characters.

17 Being in a tense emotional situation scares me.

18 When I see someone being treated unfairly, I sometimes don't feel very much pity for them.

19 I am usually pretty effective in dealing with emergencies.

20 I am often quite touched by things that I see happen.

21 I believe that there are two sides to every question and try to look at them both.

22 I would describe myself as a pretty soft-hearted person.

23 When I watch a good movie, I can very easily put myself in the place of a leading character

24 I tend to lose control during emergencies.

25 When I'm upset at someone, I usually try to "put myself in his shoes" for a while.

26 When I am reading an interesting story or novel, I imagine how I would feel if the events in the story were happening to me.

27 When I see someone who badly needs help in an emergency, I go to pieces.

28 Before criticizing somebody, I try to imagine how I would feel if I were in their place.

gender "M" (male) or "F" (female)

Note

There are four domains

Fantasy: items 1, 5, 7, 12, 16, 23, 26

Perspective taking: items 3, 8, 11, 15, 21, 25, 28

Empathic concern: items 2, 4, 9, 14, 18, 20, 22

Personal distress: items 6, 10, 13, 17, 19, 24, 27,

References

Briganti, G., Kempenaers, C., Braun, S., Fried, E. I., & Linkowski, P. (2018). Network analysis of empathy items from the interpersonal reactivity index in 1973 young adults. Psychiatry research, 265, 87-92.

Examples

data("iri")

# labels
iri_labels <- BGGM:::iri_labels