Data: Resilience Scale of Adults (RSA)

A dataset containing items from the Resilience Scale of Adults (RSA). There are 33 items and 675 observations

data("rsa")

Format

A data frame with 28 variables and 1973 observations (5 point Likert scale)

Details

1 My plans for the future are

2 When something unforeseen happens

3 My family understanding of what is important in life is

4 I feel that my future looks

5 My goals

6 I can discuss personal issues with

7 I feel

8 I enjoy being

9 Those who are good at encouraging are

10 The bonds among my friends

11 My personal problems

12 When a family member experiences a crisis/emergency

13 My family is characterised by

14 To be flexible in social settings

15 I get support from

16 In difficult periods my family

17 My judgements and decisions

18 New friendships are something

19 When needed, I have

20 I am at my best when I

21 Meeting new people is

22 When I am with others

23 When I start on new things/projects

24 Facing other people, our family acts

25 Belief in myself

26 For me, thinking of good topics of conversation is

27 My close friends/family members

28 I am good at

29 In my family, we like to

30 Rules and regular routines

31 In difficult periods I have a tendency to

32 My goals for the future are

33 Events in my life that I cannot influence

gender "M" (male) or "F" (female)

Note

There are 6 domains

Planned future: items 1, 4, 5, 32

Perception of self: items 2, 11, 17, 25, 31, 33

Family cohesion: items 3, 7, 13, 16, 24, 29

Social resources: items 6, 9, 10, 12, 15, 19, 27

Social Competence: items 8, 14, 18, 21, 22, 26,

Structured style: items 23, 28, 30

References

Briganti, G., & Linkowski, P. (2019). Item and domain network structures of the Resilience Scale for Adults in 675 university students. Epidemiology and psychiatric sciences, 1-9.

Examples

data("rsa")

# labels
rsa_labels <- BGGM:::rsa_labels