Summary Method for var_estimate Objects

Summarize the posterior distribution of each partial correlation and regression coefficient with the posterior mean, standard deviation, and credible intervals.

Usage

# S3 method for class 'var_estimate'
summary(object, cred = 0.95, ...)

Arguments

object

An object of class var_estimate

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored.

Value

A dataframe containing the summarized posterior distributions, including both the partial correlations and the regression coefficients.

pcor_results A data frame including the summarized partial correlations

beta_results A list containing the summarized regression coefficients (one data frame for each outcome)

See also

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# fit model with alias (var_estimate also works)
fit <- var_estimate(Y, progress = FALSE)

# summary ('pcor')
print(
summary(fit, cred = 0.95),
param = "pcor",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Partial Correlations: 
#> 
#>                   Relation Post.mean Post.sd Cred.lb Cred.ub
#>  interested--disinterested    -0.179   0.103  -0.372   0.036
#>        interested--excited     0.389   0.089   0.201   0.548
#>     disinterested--excited    -0.158   0.107  -0.369   0.050
#>          interested--upset    -0.209   0.097  -0.396  -0.016
#>       disinterested--upset    -0.027   0.111  -0.240   0.195
#>             excited--upset    -0.127   0.103  -0.328   0.076
#>         interested--strong     0.329   0.095   0.135   0.502
#>      disinterested--strong     0.094   0.104  -0.115   0.284
#>            excited--strong     0.487   0.083   0.312   0.638
#>              upset--strong     0.118   0.100  -0.078   0.313
#>       interested--stressed     0.279   0.093   0.085   0.454
#>    disinterested--stressed     0.150   0.101  -0.042   0.348
#>          excited--stressed    -0.176   0.105  -0.379   0.037
#>            upset--stressed     0.353   0.087   0.170   0.515
#>           strong--stressed    -0.019   0.104  -0.228   0.182
#>          interested--steps     0.071   0.098  -0.125   0.263
#>       disinterested--steps    -0.083   0.107  -0.291   0.128
#>             excited--steps     0.000   0.105  -0.212   0.194
#>               upset--steps    -0.042   0.102  -0.233   0.165
#>              strong--steps     0.178   0.101  -0.026   0.369
#>            stressed--steps    -0.022   0.112  -0.240   0.199
#> --- 
#> 

# summary ('beta')
print(
summary(fit, cred = 0.95),
param = "beta",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Coefficients: 
#> 
#> interested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.224   0.179  -0.128   0.572
#>  disinterested.l1    -0.049   0.123  -0.294   0.194
#>        excited.l1    -0.082   0.196  -0.457   0.303
#>          upset.l1    -0.154   0.128  -0.416   0.096
#>         strong.l1     0.023   0.176  -0.328   0.370
#>       stressed.l1    -0.020   0.120  -0.257   0.221
#>          steps.l1    -0.153   0.112  -0.375   0.063
#> ---
#> disinterested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.018   0.178  -0.367   0.330
#>  disinterested.l1    -0.004   0.121  -0.238   0.233
#>        excited.l1    -0.177   0.197  -0.561   0.212
#>          upset.l1     0.257   0.130   0.009   0.511
#>         strong.l1     0.173   0.174  -0.178   0.514
#>       stressed.l1    -0.008   0.121  -0.245   0.235
#>          steps.l1     0.180   0.113  -0.040   0.403
#> ---
#> excited 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.183   0.185  -0.173   0.540
#>  disinterested.l1     0.056   0.126  -0.193   0.305
#>        excited.l1     0.004   0.203  -0.390   0.398
#>          upset.l1    -0.095   0.132  -0.363   0.158
#>         strong.l1     0.021   0.181  -0.329   0.382
#>       stressed.l1    -0.034   0.123  -0.277   0.213
#>          steps.l1    -0.206   0.115  -0.440   0.023
#> ---
#> upset 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.104   0.171  -0.443   0.229
#>  disinterested.l1    -0.018   0.116  -0.242   0.209
#>        excited.l1     0.052   0.189  -0.315   0.421
#>          upset.l1     0.431   0.122   0.187   0.675
#>         strong.l1     0.052   0.170  -0.273   0.387
#>       stressed.l1    -0.042   0.116  -0.269   0.183
#>          steps.l1     0.150   0.108  -0.057   0.361
#> ---
#> strong 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.177   0.187  -0.185   0.556
#>  disinterested.l1     0.051   0.126  -0.196   0.293
#>        excited.l1    -0.083   0.203  -0.485   0.304
#>          upset.l1     0.057   0.135  -0.216   0.314
#>         strong.l1     0.180   0.182  -0.182   0.534
#>       stressed.l1    -0.077   0.124  -0.318   0.169
#>          steps.l1    -0.090   0.115  -0.317   0.135
#> ---
#> stressed 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.013   0.172  -0.329   0.354
#>  disinterested.l1     0.090   0.117  -0.143   0.316
#>        excited.l1     0.089   0.190  -0.292   0.457
#>          upset.l1     0.316   0.121   0.074   0.554
#>         strong.l1    -0.066   0.167  -0.389   0.258
#>       stressed.l1     0.155   0.116  -0.076   0.382
#>          steps.l1     0.201   0.108  -0.008   0.414
#> ---
#> steps 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.113   0.187  -0.249   0.474
#>  disinterested.l1    -0.026   0.126  -0.273   0.223
#>        excited.l1     0.096   0.203  -0.291   0.497
#>          upset.l1    -0.088   0.133  -0.346   0.178
#>         strong.l1    -0.186   0.183  -0.546   0.173
#>       stressed.l1     0.128   0.125  -0.115   0.370
#>          steps.l1     0.041   0.117  -0.192   0.270
#> ---

# }