Summary Method for var_estimate Objects

Summarize the posterior distribution of each partial correlation and regression coefficient with the posterior mean, standard deviation, and credible intervals.

Usage

# S3 method for class 'var_estimate'
summary(object, cred = 0.95, ...)

Arguments

object

An object of class var_estimate

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored.

Value

A dataframe containing the summarized posterior distributions, including both the partial correlations and the regression coefficients.

pcor_results A data frame including the summarized partial correlations

beta_results A list containing the summarized regression coefficients (one data frame for each outcome)

See also

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# fit model with alias (var_estimate also works)
fit <- var_estimate(Y, progress = FALSE)

# summary ('pcor')
print(
summary(fit, cred = 0.95),
param = "pcor",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Partial Correlations: 
#> 
#>                   Relation Post.mean Post.sd Cred.lb Cred.ub
#>  interested--disinterested    -0.180   0.102  -0.370   0.027
#>        interested--excited     0.395   0.085   0.211   0.545
#>     disinterested--excited    -0.167   0.106  -0.373   0.050
#>          interested--upset    -0.210   0.097  -0.393  -0.010
#>       disinterested--upset    -0.021   0.107  -0.229   0.190
#>             excited--upset    -0.118   0.097  -0.303   0.073
#>         interested--strong     0.314   0.088   0.134   0.483
#>      disinterested--strong     0.097   0.105  -0.112   0.297
#>            excited--strong     0.493   0.077   0.334   0.639
#>              upset--strong     0.106   0.099  -0.092   0.291
#>       interested--stressed     0.288   0.097   0.086   0.477
#>    disinterested--stressed     0.156   0.104  -0.044   0.352
#>          excited--stressed    -0.178   0.103  -0.370   0.034
#>            upset--stressed     0.353   0.094   0.153   0.531
#>           strong--stressed    -0.010   0.104  -0.218   0.199
#>          interested--steps     0.078   0.111  -0.151   0.292
#>       disinterested--steps    -0.085   0.101  -0.273   0.122
#>             excited--steps    -0.010   0.103  -0.211   0.191
#>               upset--steps    -0.048   0.100  -0.253   0.148
#>              strong--steps     0.181   0.095  -0.012   0.366
#>            stressed--steps    -0.020   0.102  -0.220   0.180
#> --- 
#> 

# summary ('beta')
print(
summary(fit, cred = 0.95),
param = "beta",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Coefficients: 
#> 
#> interested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.222   0.178  -0.129   0.565
#>  disinterested.l1    -0.050   0.123  -0.291   0.195
#>        excited.l1    -0.080   0.197  -0.470   0.311
#>          upset.l1    -0.154   0.131  -0.410   0.101
#>         strong.l1     0.026   0.180  -0.324   0.382
#>       stressed.l1    -0.019   0.121  -0.257   0.216
#>          steps.l1    -0.154   0.112  -0.373   0.065
#> ---
#> disinterested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.010   0.177  -0.363   0.340
#>  disinterested.l1    -0.005   0.122  -0.241   0.232
#>        excited.l1    -0.185   0.195  -0.563   0.207
#>          upset.l1     0.259   0.127   0.003   0.506
#>         strong.l1     0.173   0.175  -0.177   0.529
#>       stressed.l1    -0.012   0.123  -0.258   0.223
#>          steps.l1     0.177   0.112  -0.042   0.393
#> ---
#> excited 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.178   0.183  -0.180   0.548
#>  disinterested.l1     0.056   0.125  -0.185   0.310
#>        excited.l1     0.006   0.199  -0.396   0.391
#>          upset.l1    -0.097   0.131  -0.353   0.161
#>         strong.l1     0.028   0.180  -0.321   0.385
#>       stressed.l1    -0.030   0.124  -0.277   0.221
#>          steps.l1    -0.208   0.115  -0.434   0.018
#> ---
#> upset 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.097   0.171  -0.442   0.240
#>  disinterested.l1    -0.017   0.119  -0.253   0.218
#>        excited.l1     0.054   0.185  -0.315   0.418
#>          upset.l1     0.431   0.123   0.193   0.673
#>         strong.l1     0.047   0.166  -0.280   0.370
#>       stressed.l1    -0.045   0.117  -0.273   0.181
#>          steps.l1     0.152   0.108  -0.060   0.368
#> ---
#> strong 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.174   0.182  -0.180   0.538
#>  disinterested.l1     0.050   0.124  -0.186   0.294
#>        excited.l1    -0.081   0.198  -0.472   0.298
#>          upset.l1     0.055   0.134  -0.200   0.316
#>         strong.l1     0.184   0.179  -0.170   0.537
#>       stressed.l1    -0.072   0.122  -0.314   0.167
#>          steps.l1    -0.091   0.115  -0.319   0.132
#> ---
#> stressed 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.014   0.169  -0.324   0.339
#>  disinterested.l1     0.089   0.117  -0.143   0.319
#>        excited.l1     0.086   0.188  -0.286   0.459
#>          upset.l1     0.318   0.122   0.079   0.557
#>         strong.l1    -0.067   0.169  -0.407   0.263
#>       stressed.l1     0.151   0.116  -0.078   0.379
#>          steps.l1     0.205   0.108  -0.006   0.420
#> ---
#> steps 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.110   0.180  -0.237   0.474
#>  disinterested.l1    -0.024   0.126  -0.266   0.223
#>        excited.l1     0.103   0.198  -0.293   0.477
#>          upset.l1    -0.094   0.130  -0.349   0.166
#>         strong.l1    -0.185   0.181  -0.542   0.169
#>       stressed.l1     0.131   0.123  -0.109   0.379
#>          steps.l1     0.041   0.114  -0.189   0.261
#> ---

# }