Summary Method for var_estimate Objects

Summarize the posterior distribution of each partial correlation and regression coefficient with the posterior mean, standard deviation, and credible intervals.

Usage

# S3 method for class 'var_estimate'
summary(object, cred = 0.95, ...)

Arguments

object

An object of class var_estimate

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored.

Value

A dataframe containing the summarized posterior distributions, including both the partial correlations and the regression coefficients.

pcor_results A data frame including the summarized partial correlations

beta_results A list containing the summarized regression coefficients (one data frame for each outcome)

See also

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# fit model with alias (var_estimate also works)
fit <- var_estimate(Y, progress = FALSE)

# summary ('pcor')
print(
summary(fit, cred = 0.95),
param = "pcor",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Partial Correlations: 
#> 
#>                   Relation Post.mean Post.sd Cred.lb Cred.ub
#>  interested--disinterested    -0.178   0.105  -0.373   0.033
#>        interested--excited     0.388   0.088   0.207   0.553
#>     disinterested--excited    -0.168   0.105  -0.367   0.041
#>          interested--upset    -0.191   0.096  -0.379  -0.004
#>       disinterested--upset    -0.031   0.100  -0.221   0.174
#>             excited--upset    -0.136   0.105  -0.329   0.079
#>         interested--strong     0.338   0.089   0.145   0.497
#>      disinterested--strong     0.099   0.108  -0.097   0.315
#>            excited--strong     0.482   0.081   0.311   0.627
#>              upset--strong     0.116   0.104  -0.096   0.311
#>       interested--stressed     0.271   0.095   0.084   0.447
#>    disinterested--stressed     0.159   0.100  -0.043   0.351
#>          excited--stressed    -0.156   0.106  -0.355   0.053
#>            upset--stressed     0.348   0.095   0.161   0.523
#>           strong--stressed    -0.027   0.099  -0.219   0.160
#>          interested--steps     0.071   0.105  -0.147   0.271
#>       disinterested--steps    -0.093   0.107  -0.299   0.124
#>             excited--steps    -0.002   0.110  -0.222   0.207
#>               upset--steps    -0.056   0.108  -0.270   0.158
#>              strong--steps     0.177   0.105  -0.032   0.384
#>            stressed--steps    -0.016   0.101  -0.228   0.179
#> --- 
#> 

# summary ('beta')
print(
summary(fit, cred = 0.95),
param = "beta",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Coefficients: 
#> 
#> interested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.218   0.183  -0.139   0.584
#>  disinterested.l1    -0.050   0.124  -0.294   0.196
#>        excited.l1    -0.078   0.195  -0.455   0.308
#>          upset.l1    -0.154   0.131  -0.410   0.108
#>         strong.l1     0.023   0.175  -0.322   0.373
#>       stressed.l1    -0.018   0.122  -0.267   0.219
#>          steps.l1    -0.153   0.114  -0.372   0.069
#> ---
#> disinterested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.011   0.179  -0.360   0.340
#>  disinterested.l1    -0.007   0.121  -0.244   0.235
#>        excited.l1    -0.183   0.194  -0.559   0.209
#>          upset.l1     0.258   0.129   0.006   0.514
#>         strong.l1     0.172   0.175  -0.187   0.515
#>       stressed.l1    -0.010   0.121  -0.246   0.231
#>          steps.l1     0.180   0.114  -0.041   0.400
#> ---
#> excited 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.175   0.183  -0.178   0.540
#>  disinterested.l1     0.058   0.123  -0.184   0.300
#>        excited.l1     0.008   0.198  -0.383   0.403
#>          upset.l1    -0.095   0.133  -0.352   0.169
#>         strong.l1     0.023   0.178  -0.333   0.368
#>       stressed.l1    -0.031   0.124  -0.277   0.211
#>          steps.l1    -0.208   0.115  -0.437   0.013
#> ---
#> upset 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.098   0.171  -0.428   0.248
#>  disinterested.l1    -0.017   0.117  -0.248   0.208
#>        excited.l1     0.052   0.187  -0.321   0.416
#>          upset.l1     0.426   0.122   0.186   0.665
#>         strong.l1     0.048   0.167  -0.284   0.371
#>       stressed.l1    -0.042   0.117  -0.277   0.183
#>          steps.l1     0.150   0.108  -0.066   0.359
#> ---
#> strong 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.173   0.186  -0.195   0.534
#>  disinterested.l1     0.052   0.126  -0.195   0.299
#>        excited.l1    -0.079   0.202  -0.478   0.313
#>          upset.l1     0.056   0.135  -0.208   0.328
#>         strong.l1     0.178   0.181  -0.177   0.526
#>       stressed.l1    -0.073   0.124  -0.320   0.167
#>          steps.l1    -0.090   0.117  -0.325   0.137
#> ---
#> stressed 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.013   0.173  -0.328   0.357
#>  disinterested.l1     0.089   0.115  -0.137   0.316
#>        excited.l1     0.084   0.182  -0.274   0.442
#>          upset.l1     0.312   0.123   0.071   0.554
#>         strong.l1    -0.064   0.167  -0.395   0.261
#>       stressed.l1     0.153   0.115  -0.075   0.377
#>          steps.l1     0.204   0.108  -0.005   0.413
#> ---
#> steps 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.110   0.181  -0.252   0.462
#>  disinterested.l1    -0.021   0.127  -0.268   0.235
#>        excited.l1     0.102   0.199  -0.290   0.500
#>          upset.l1    -0.094   0.131  -0.346   0.162
#>         strong.l1    -0.186   0.182  -0.541   0.173
#>       stressed.l1     0.133   0.125  -0.117   0.379
#>          steps.l1     0.038   0.116  -0.194   0.268
#> ---

# }