Summary Method for var_estimate Objects

Summarize the posterior distribution of each partial correlation and regression coefficient with the posterior mean, standard deviation, and credible intervals.

# S3 method for var_estimate
summary(object, cred = 0.95, ...)

Arguments

object
An object of class var_estimate

cred
Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...
Currently ignored.

Value

A dataframe containing the summarized posterior distributions, including both the partial correlations and the regression coefficients.

pcor_results A data frame including the summarized partial correlations

beta_results A list containing the summarized regression coefficients (one data frame for each outcome)

See also

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# fit model with alias (var_estimate also works)
fit <- var_estimate(Y, progress = FALSE)

# summary ('pcor')
print(
summary(fit, cred = 0.95),
param = "pcor",
)

#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Partial Correlations: 
#> 
#>                   Relation Post.mean Post.sd Cred.lb Cred.ub
#>  interested--disinterested    -0.186   0.098  -0.375   0.004
#>        interested--excited     0.378   0.085   0.203   0.536
#>     disinterested--excited    -0.169   0.098  -0.367   0.022
#>          interested--upset    -0.196   0.104  -0.404   0.008
#>       disinterested--upset    -0.024   0.101  -0.225   0.169
#>             excited--upset    -0.134   0.108  -0.334   0.085
#>         interested--strong     0.328   0.094   0.139   0.503
#>      disinterested--strong     0.102   0.106  -0.102   0.303
#>            excited--strong     0.493   0.077   0.333   0.631
#>              upset--strong     0.113   0.105  -0.094   0.313
#>       interested--stressed     0.275   0.097   0.083   0.455
#>    disinterested--stressed     0.162   0.100  -0.038   0.349
#>          excited--stressed    -0.175   0.100  -0.366   0.015
#>            upset--stressed     0.344   0.098   0.130   0.523
#>           strong--stressed    -0.013   0.100  -0.204   0.186
#>          interested--steps     0.074   0.099  -0.118   0.267
#>       disinterested--steps    -0.079   0.105  -0.283   0.127
#>             excited--steps    -0.008   0.107  -0.224   0.203
#>               upset--steps    -0.038   0.105  -0.235   0.171
#>              strong--steps     0.184   0.101  -0.016   0.381
#>            stressed--steps    -0.011   0.102  -0.216   0.187
#> --- 
#> 

# summary ('beta')
print(
summary(fit, cred = 0.95),
param = "beta",
)

#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Coefficients: 
#> 
#> interested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.220   0.181  -0.134   0.576
#>  disinterested.l1    -0.050   0.121  -0.283   0.185
#>        excited.l1    -0.084   0.198  -0.477   0.308
#>          upset.l1    -0.153   0.128  -0.404   0.097
#>         strong.l1     0.029   0.178  -0.327   0.382
#>       stressed.l1    -0.020   0.119  -0.257   0.210
#>          steps.l1    -0.153   0.114  -0.374   0.066
#> ---
#> disinterested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.013   0.178  -0.365   0.330
#>  disinterested.l1    -0.004   0.121  -0.243   0.235
#>        excited.l1    -0.184   0.197  -0.567   0.209
#>          upset.l1     0.256   0.128   0.009   0.509
#>         strong.l1     0.173   0.177  -0.164   0.525
#>       stressed.l1    -0.009   0.120  -0.240   0.225
#>          steps.l1     0.181   0.112  -0.038   0.400
#> ---
#> excited 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.176   0.186  -0.192   0.533
#>  disinterested.l1     0.056   0.126  -0.199   0.307
#>        excited.l1    -0.001   0.205  -0.413   0.398
#>          upset.l1    -0.099   0.132  -0.352   0.161
#>         strong.l1     0.030   0.183  -0.322   0.390
#>       stressed.l1    -0.029   0.123  -0.273   0.209
#>          steps.l1    -0.205   0.115  -0.424   0.019
#> ---
#> upset 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.096   0.173  -0.428   0.245
#>  disinterested.l1    -0.018   0.115  -0.247   0.207
#>        excited.l1     0.055   0.189  -0.317   0.426
#>          upset.l1     0.432   0.124   0.190   0.680
#>         strong.l1     0.046   0.170  -0.291   0.374
#>       stressed.l1    -0.044   0.117  -0.277   0.190
#>          steps.l1     0.150   0.109  -0.065   0.365
#> ---
#> strong 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.171   0.187  -0.193   0.538
#>  disinterested.l1     0.049   0.126  -0.206   0.299
#>        excited.l1    -0.085   0.202  -0.478   0.316
#>          upset.l1     0.054   0.132  -0.205   0.303
#>         strong.l1     0.185   0.183  -0.177   0.544
#>       stressed.l1    -0.073   0.122  -0.310   0.162
#>          steps.l1    -0.089   0.116  -0.317   0.136
#> ---
#> stressed 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.015   0.173  -0.326   0.347
#>  disinterested.l1     0.089   0.119  -0.137   0.325
#>        excited.l1     0.085   0.188  -0.284   0.450
#>          upset.l1     0.317   0.124   0.073   0.557
#>         strong.l1    -0.065   0.169  -0.400   0.256
#>       stressed.l1     0.150   0.114  -0.070   0.371
#>          steps.l1     0.205   0.108  -0.008   0.415
#> ---
#> steps 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.106   0.184  -0.250   0.467
#>  disinterested.l1    -0.023   0.125  -0.266   0.222
#>        excited.l1     0.099   0.201  -0.299   0.489
#>          upset.l1    -0.095   0.134  -0.356   0.163
#>         strong.l1    -0.182   0.180  -0.536   0.179
#>       stressed.l1     0.133   0.122  -0.100   0.374
#>          steps.l1     0.039   0.117  -0.190   0.269
#> ---

# }