Summary Method for var_estimate Objects

Summarize the posterior distribution of each partial correlation and regression coefficient with the posterior mean, standard deviation, and credible intervals.

Usage

# S3 method for class 'var_estimate'
summary(object, cred = 0.95, ...)

Arguments

object

An object of class var_estimate

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored.

Value

A dataframe containing the summarized posterior distributions, including both the partial correlations and the regression coefficients.

pcor_results A data frame including the summarized partial correlations

beta_results A list containing the summarized regression coefficients (one data frame for each outcome)

See also

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# fit model with alias (var_estimate also works)
fit <- var_estimate(Y, progress = FALSE)

# summary ('pcor')
print(
summary(fit, cred = 0.95),
param = "pcor",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Partial Correlations: 
#> 
#>                   Relation Post.mean Post.sd Cred.lb Cred.ub
#>  interested--disinterested    -0.188   0.101  -0.379   0.021
#>        interested--excited     0.379   0.084   0.202   0.532
#>     disinterested--excited    -0.171   0.095  -0.349   0.018
#>          interested--upset    -0.202   0.101  -0.392  -0.006
#>       disinterested--upset    -0.037   0.107  -0.238   0.173
#>             excited--upset    -0.138   0.101  -0.343   0.051
#>         interested--strong     0.326   0.098   0.128   0.507
#>      disinterested--strong     0.105   0.103  -0.105   0.302
#>            excited--strong     0.491   0.082   0.313   0.640
#>              upset--strong     0.119   0.103  -0.085   0.316
#>       interested--stressed     0.269   0.099   0.073   0.457
#>    disinterested--stressed     0.163   0.100  -0.032   0.347
#>          excited--stressed    -0.162   0.105  -0.360   0.051
#>            upset--stressed     0.354   0.095   0.159   0.538
#>           strong--stressed    -0.011   0.112  -0.231   0.202
#>          interested--steps     0.078   0.107  -0.130   0.290
#>       disinterested--steps    -0.076   0.101  -0.273   0.118
#>             excited--steps    -0.005   0.104  -0.208   0.202
#>               upset--steps    -0.057   0.107  -0.272   0.141
#>              strong--steps     0.178   0.099  -0.018   0.381
#>            stressed--steps    -0.016   0.108  -0.226   0.200
#> --- 
#> 

# summary ('beta')
print(
summary(fit, cred = 0.95),
param = "beta",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Coefficients: 
#> 
#> interested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.220   0.178  -0.134   0.571
#>  disinterested.l1    -0.049   0.124  -0.287   0.199
#>        excited.l1    -0.074   0.195  -0.467   0.308
#>          upset.l1    -0.150   0.128  -0.398   0.103
#>         strong.l1     0.023   0.172  -0.309   0.351
#>       stressed.l1    -0.016   0.123  -0.259   0.226
#>          steps.l1    -0.154   0.114  -0.378   0.066
#> ---
#> disinterested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.013   0.180  -0.377   0.335
#>  disinterested.l1    -0.005   0.124  -0.245   0.236
#>        excited.l1    -0.183   0.195  -0.572   0.207
#>          upset.l1     0.255   0.126   0.010   0.507
#>         strong.l1     0.171   0.174  -0.167   0.530
#>       stressed.l1    -0.010   0.119  -0.237   0.223
#>          steps.l1     0.181   0.113  -0.034   0.403
#> ---
#> excited 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.176   0.180  -0.180   0.527
#>  disinterested.l1     0.056   0.126  -0.197   0.302
#>        excited.l1     0.010   0.197  -0.376   0.397
#>          upset.l1    -0.094   0.130  -0.351   0.153
#>         strong.l1     0.025   0.178  -0.321   0.387
#>       stressed.l1    -0.028   0.125  -0.276   0.213
#>          steps.l1    -0.208   0.116  -0.436   0.020
#> ---
#> upset 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.095   0.172  -0.440   0.235
#>  disinterested.l1    -0.022   0.118  -0.251   0.212
#>        excited.l1     0.048   0.188  -0.316   0.422
#>          upset.l1     0.430   0.125   0.188   0.670
#>         strong.l1     0.049   0.168  -0.283   0.380
#>       stressed.l1    -0.044   0.119  -0.277   0.188
#>          steps.l1     0.151   0.109  -0.065   0.362
#> ---
#> strong 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.173   0.184  -0.187   0.531
#>  disinterested.l1     0.049   0.124  -0.199   0.294
#>        excited.l1    -0.077   0.201  -0.470   0.326
#>          upset.l1     0.059   0.132  -0.200   0.318
#>         strong.l1     0.182   0.180  -0.170   0.537
#>       stressed.l1    -0.070   0.125  -0.317   0.182
#>          steps.l1    -0.091   0.116  -0.320   0.137
#> ---
#> stressed 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.012   0.173  -0.327   0.357
#>  disinterested.l1     0.090   0.117  -0.139   0.323
#>        excited.l1     0.091   0.187  -0.279   0.454
#>          upset.l1     0.313   0.126   0.066   0.562
#>         strong.l1    -0.072   0.168  -0.401   0.260
#>       stressed.l1     0.154   0.118  -0.076   0.382
#>          steps.l1     0.205   0.110  -0.009   0.421
#> ---
#> steps 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.106   0.183  -0.269   0.461
#>  disinterested.l1    -0.024   0.126  -0.271   0.225
#>        excited.l1     0.104   0.200  -0.294   0.495
#>          upset.l1    -0.089   0.132  -0.348   0.163
#>         strong.l1    -0.183   0.177  -0.534   0.163
#>       stressed.l1     0.131   0.126  -0.124   0.384
#>          steps.l1     0.041   0.116  -0.184   0.273
#> ---

# }