Summary Method for var_estimate Objects

Summarize the posterior distribution of each partial correlation and regression coefficient with the posterior mean, standard deviation, and credible intervals.

Usage

# S3 method for class 'var_estimate'
summary(object, cred = 0.95, ...)

Arguments

object

An object of class var_estimate

cred

Numeric. The credible interval width for summarizing the posterior distributions (defaults to 0.95; must be between 0 and 1).

...

Currently ignored.

Value

A dataframe containing the summarized posterior distributions, including both the partial correlations and the regression coefficients.

pcor_results A data frame including the summarized partial correlations

beta_results A list containing the summarized regression coefficients (one data frame for each outcome)

See also

Examples

# \donttest{
# data
Y <- subset(ifit, id == 1)[,-1]

# fit model with alias (var_estimate also works)
fit <- var_estimate(Y, progress = FALSE)

# summary ('pcor')
print(
summary(fit, cred = 0.95),
param = "pcor",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Partial Correlations: 
#> 
#>                   Relation Post.mean Post.sd Cred.lb Cred.ub
#>  interested--disinterested    -0.182   0.096  -0.365   0.007
#>        interested--excited     0.385   0.089   0.198   0.552
#>     disinterested--excited    -0.170   0.100  -0.364   0.034
#>          interested--upset    -0.199   0.099  -0.382   0.009
#>       disinterested--upset    -0.037   0.104  -0.246   0.165
#>             excited--upset    -0.145   0.103  -0.346   0.057
#>         interested--strong     0.326   0.093   0.142   0.495
#>      disinterested--strong     0.095   0.107  -0.116   0.301
#>            excited--strong     0.490   0.079   0.322   0.636
#>              upset--strong     0.122   0.106  -0.095   0.317
#>       interested--stressed     0.275   0.097   0.076   0.457
#>    disinterested--stressed     0.154   0.101  -0.062   0.343
#>          excited--stressed    -0.166   0.099  -0.352   0.039
#>            upset--stressed     0.355   0.089   0.176   0.522
#>           strong--stressed    -0.012   0.102  -0.218   0.188
#>          interested--steps     0.080   0.105  -0.128   0.276
#>       disinterested--steps    -0.084   0.099  -0.271   0.111
#>             excited--steps    -0.013   0.107  -0.224   0.196
#>               upset--steps    -0.052   0.108  -0.266   0.150
#>              strong--steps     0.182   0.101  -0.027   0.370
#>            stressed--steps    -0.015   0.104  -0.221   0.188
#> --- 
#> 

# summary ('beta')
print(
summary(fit, cred = 0.95),
param = "beta",
)
#> BGGM: Bayesian Gaussian Graphical Models 
#> --- 
#> Vector Autoregressive Model (VAR) 
#> --- 
#> Coefficients: 
#> 
#> interested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.218   0.179  -0.131   0.565
#>  disinterested.l1    -0.048   0.122  -0.290   0.186
#>        excited.l1    -0.074   0.196  -0.453   0.308
#>          upset.l1    -0.152   0.128  -0.409   0.095
#>         strong.l1     0.024   0.175  -0.321   0.365
#>       stressed.l1    -0.019   0.122  -0.256   0.221
#>          steps.l1    -0.155   0.114  -0.370   0.066
#> ---
#> disinterested 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.012   0.178  -0.360   0.330
#>  disinterested.l1    -0.006   0.122  -0.251   0.233
#>        excited.l1    -0.185   0.191  -0.559   0.193
#>          upset.l1     0.259   0.128   0.007   0.509
#>         strong.l1     0.174   0.175  -0.164   0.525
#>       stressed.l1    -0.009   0.122  -0.247   0.231
#>          steps.l1     0.182   0.110  -0.025   0.404
#> ---
#> excited 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.176   0.180  -0.182   0.529
#>  disinterested.l1     0.058   0.124  -0.190   0.288
#>        excited.l1     0.010   0.199  -0.387   0.388
#>          upset.l1    -0.095   0.132  -0.351   0.159
#>         strong.l1     0.024   0.176  -0.326   0.366
#>       stressed.l1    -0.031   0.123  -0.277   0.216
#>          steps.l1    -0.209   0.115  -0.435   0.016
#> ---
#> upset 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1    -0.097   0.174  -0.431   0.238
#>  disinterested.l1    -0.017   0.119  -0.246   0.220
#>        excited.l1     0.054   0.190  -0.320   0.422
#>          upset.l1     0.430   0.121   0.193   0.667
#>         strong.l1     0.046   0.169  -0.273   0.374
#>       stressed.l1    -0.043   0.113  -0.268   0.182
#>          steps.l1     0.153   0.110  -0.056   0.368
#> ---
#> strong 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.171   0.179  -0.184   0.527
#>  disinterested.l1     0.052   0.125  -0.192   0.295
#>        excited.l1    -0.074   0.197  -0.463   0.319
#>          upset.l1     0.056   0.132  -0.203   0.309
#>         strong.l1     0.179   0.178  -0.164   0.531
#>       stressed.l1    -0.074   0.125  -0.323   0.174
#>          steps.l1    -0.091   0.117  -0.323   0.136
#> ---
#> stressed 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.016   0.174  -0.319   0.368
#>  disinterested.l1     0.091   0.116  -0.138   0.318
#>        excited.l1     0.089   0.186  -0.274   0.457
#>          upset.l1     0.316   0.122   0.073   0.558
#>         strong.l1    -0.069   0.169  -0.404   0.266
#>       stressed.l1     0.154   0.113  -0.069   0.376
#>          steps.l1     0.203   0.106  -0.001   0.416
#> ---
#> steps 
#> 
#>          Relation Post.mean Post.sd Cred.lb Cred.ub
#>     interested.l1     0.105   0.184  -0.253   0.459
#>  disinterested.l1    -0.022   0.128  -0.271   0.229
#>        excited.l1     0.103   0.205  -0.305   0.516
#>          upset.l1    -0.092   0.133  -0.358   0.173
#>         strong.l1    -0.180   0.182  -0.527   0.174
#>       stressed.l1     0.130   0.125  -0.112   0.374
#>          steps.l1     0.041   0.118  -0.195   0.270
#> ---

# }